### 2024-05-30 #6814 @editorialbot
[PRE REVIEW]: pathways: life cycle assessment of energy transition scenarios
**Submitting author:** @romainsacchi (<a href="http://orcid.org/0000-0003-1440-0905">Romain Sacchi</a>)
**Repository:** https://github.com/polca/pathways
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.1
