### 2024-05-30 #6814 @kyleniemeyer
Hello @romainsacchi, unfortunately we do not have any editors available to handle your submission right now, so I have to put this on our waitlist until someone frees up.

In the meantime, reviewer recommendations would be welcome (just please don't formally tag anyone with @).

In addition, to fix the paper compilation issue, please put quotations around the affiliation field with multiple numbers: `affiliation: "1,2"`
