+++ setup.py
@@ -53,6 +53,7 @@ setup(
         "datapackage",
         "pyprind",
         "platformdirs",
+        "fs",
     ],
     url="https://github.com/polca/premise",
     description="Scenario-level LCA of energy systems and transition pathways",
