+++ requirements.txt
@@ -10,3 +10,4 @@ bw2calc == 2.0.DEV16
 datapackage
 pyprind
 platformdirs
+fs
