### 2024-06-13 3e1368d
Fix issues with fetching model name in datapackage.json
