+++ pathways/utils.py
@@ -10,19 +10,20 @@ dictionary, checking unclassified activities, and getting activity indices.
 
 import csv
 import logging
+from datetime import datetime
 from pathlib import Path
 from typing import Any, Dict, List, Set, Tuple, Union
 
 import numpy as np
 import xarray as xr
 import yaml
+import pandas as pd
 from premise.geomap import Geomap
 
 from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, USER_LOGS_DIR
 
 CLASSIFICATIONS = DATA_DIR / "activities_classifications.yaml"
 UNITS_CONVERSION = DATA_DIR / "units_conversion.yaml"
-
 
 logging.basicConfig(
     level=logging.DEBUG,
@@ -105,8 +106,8 @@ def harmonize_units(scenario: xr.DataArray, variables: list) -> xr.DataArray:
             )
             # multiply scenario by conversion factors
             scenario.loc[dict(variables=variables)] *= conversion_factors[
-                :, np.newaxis, np.newaxis
-            ]
+                                                       :, np.newaxis, np.newaxis
+                                                       ]
             # update units
             scenario.attrs["units"] = {var: "EJ/yr" for var in variables}
 
@@ -114,7 +115,7 @@ def harmonize_units(scenario: xr.DataArray, variables: list) -> xr.DataArray:
 
 
 def get_unit_conversion_factors(
-    scenario_unit: dict, dataset_unit: list, unit_mapping: dict
+        scenario_unit: dict, dataset_unit: list, unit_mapping: dict
 ) -> np.ndarray:
     """
     Get the unit conversion factors for a given scenario unit and dataset unit.
@@ -139,15 +140,15 @@ def load_units_conversion() -> dict:
 
 
 def create_lca_results_array(
-    methods: [List[str], None],
-    years: List[int],
-    regions: List[str],
-    locations: List[str],
-    models: List[str],
-    scenarios: List[str],
-    classifications: dict,
-    mapping: dict,
-    use_distributions: bool = False,
+        methods: [List[str], None],
+        years: List[int],
+        regions: List[str],
+        locations: List[str],
+        models: List[str],
+        scenarios: List[str],
+        classifications: dict,
+        mapping: dict,
+        use_distributions: bool = False,
 ) -> xr.DataArray:
     """
     Create an xarray DataArray to store Life Cycle Assessment (LCA) results.
@@ -216,10 +217,47 @@ def create_lca_results_array(
     return xr.DataArray(np.zeros(dims), coords=coords, dims=list(coords.keys()))
 
 
+def export_results_to_parquet(lca_results: xr.DataArray, filepath: str):
+    """
+    Export the LCA results to a parquet file.
+    :param lca_results: Xarray DataArray with LCA results.
+    :param filepath: The path to the parquet file.
+    :return: None
+    """
+    if filepath is None:
+        filepath = f"results_{datetime.now().strftime('%Y%m%d_%H%M%S')}.gzip"
+    else:
+        filepath = f"{filepath}.gzip"
+
+    flattened_data = lca_results.values.flatten()
+
+    # Step 2: Find the indices of non-zero values
+    non_zero_indices = np.nonzero(flattened_data)[0]
+
+    # Step 3: Extract non-zero values
+    non_zero_values = flattened_data[non_zero_indices]
+
+    # Step 4: Get the shape of the original DataArray
+    original_shape = lca_results.shape
+
+    # Step 5: Find the coordinates corresponding to the non-zero indices
+    coords = np.array(np.unravel_index(non_zero_indices, original_shape)).T
+
[…]
