+++ pathways/utils.py
@@ -15,9 +15,9 @@ from pathlib import Path
 from typing import Any, Dict, List, Set, Tuple, Union
 
 import numpy as np
+import pandas as pd
 import xarray as xr
 import yaml
-import pandas as pd
 from premise.geomap import Geomap
 
 from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, USER_LOGS_DIR
@@ -106,8 +106,8 @@ def harmonize_units(scenario: xr.DataArray, variables: list) -> xr.DataArray:
             )
             # multiply scenario by conversion factors
             scenario.loc[dict(variables=variables)] *= conversion_factors[
-                                                       :, np.newaxis, np.newaxis
-                                                       ]
+                :, np.newaxis, np.newaxis
+            ]
             # update units
             scenario.attrs["units"] = {var: "EJ/yr" for var in variables}
 
@@ -115,7 +115,7 @@ def harmonize_units(scenario: xr.DataArray, variables: list) -> xr.DataArray:
 
 
 def get_unit_conversion_factors(
-        scenario_unit: dict, dataset_unit: list, unit_mapping: dict
+    scenario_unit: dict, dataset_unit: list, unit_mapping: dict
 ) -> np.ndarray:
     """
     Get the unit conversion factors for a given scenario unit and dataset unit.
@@ -140,15 +140,15 @@ def load_units_conversion() -> dict:
 
 
 def create_lca_results_array(
-        methods: [List[str], None],
-        years: List[int],
-        regions: List[str],
-        locations: List[str],
-        models: List[str],
-        scenarios: List[str],
-        classifications: dict,
-        mapping: dict,
-        use_distributions: bool = False,
+    methods: [List[str], None],
+    years: List[int],
+    regions: List[str],
+    locations: List[str],
+    models: List[str],
+    scenarios: List[str],
+    classifications: dict,
+    mapping: dict,
+    use_distributions: bool = False,
 ) -> xr.DataArray:
     """
     Create an xarray DataArray to store Life Cycle Assessment (LCA) results.
@@ -245,19 +245,22 @@ def export_results_to_parquet(lca_results: xr.DataArray, filepath: str):
 
     # Step 6: Create a pandas DataFrame with non-zero values and corresponding coordinates
     coord_names = list(lca_results.dims)
-    coord_values = {dim: lca_results.coords[dim].values[coords[:, i]] for i, dim in enumerate(coord_names)}
+    coord_values = {
+        dim: lca_results.coords[dim].values[coords[:, i]]
+        for i, dim in enumerate(coord_names)
+    }
 
     df = pd.DataFrame(coord_values)
-    df['value'] = non_zero_values
-    df.to_parquet(path=filepath, compression='gzip')
+    df["value"] = non_zero_values
+    df.to_parquet(path=filepath, compression="gzip")
 
     print(f"Results exported to {filepath}")
 
 
 def display_results(
-        lca_results: Union[xr.DataArray, None],
-        cutoff: float = 0.001,
-        interpolate: bool = False,
+    lca_results: Union[xr.DataArray, None],
+    cutoff: float = 0.001,
+    interpolate: bool = False,
 ) -> xr.DataArray:
     """
     Display the LCA results.
@@ -328,12 +331,12 @@ def clean_cache_directory():
 
 
 def resize_scenario_data(
-        scenario_data: xr.DataArray,
-        model: List[str],
-        scenario: List[str],
-        region: List[str],
-        year: List[int],
-        variables: List[str],
+    scenario_data: xr.DataArray,
+    model: List[str],
+    scenario: List[str],
+    region: List[str],
+    year: List[int],
[…]
