+++ pathways/lca.py
@@ -43,6 +43,7 @@ from .utils import (
     get_unit_conversion_factors,
     read_categories_from_yaml,
     read_indices_csv,
+    get_all_indices,
 )
 
 logging.basicConfig(
@@ -184,43 +185,102 @@ def find_uncertain_parameters(
 
     return uncertain_parameters
 
-
-def remove_double_accounting(
-    lca: bc.LCA,
-    demand: Dict,
-    activities_to_exclude: List[int],
-    exceptions: List[int],
+def remove_double_counting(
+    technosphere_matrix: np.array,
+    activities_to_zero: List[int],
+    infra: List[int]
 ):
     """
-    Remove double counting from a technosphere matrix.
-    :param lca: bw2calc.LCA object
+    Remove double counting from a technosphere matrix by zeroing out the demanded row values
+    in all columns, except for those on the diagonal.
+    :param technosphere_matrix: bw2calc.LCA object
     :param demand: dict with demand values
-    :param activities_to_exclude: list of activities to exclude
-    :param exceptions: list of exceptions
-    :return: Technosphere matrix with double counting removed.
+    :param activities_to_exclude: list of row indices to zero out
+    :return: Technosphere matrix with double counting removed
     """
+    print("Activities to zero (rows):", activities_to_zero)
 
-    tm_original = lca.technosphere_matrix.copy()
-    tm_modified = tm_original.tocoo()
+    # Copy and convert the technosphere matrix to COO format for easy manipulation
+    # technosphere_matrix_original = technosphere_matrix.technosphere_matrix.copy()
+    technosphere_matrix = technosphere_matrix.tocoo()
 
-    for act in activities_to_exclude:
-        row_idx = np.where(tm_modified.col == act)[0]
+    # Create a mask for elements to zero out
+    mask = np.isin(technosphere_matrix.row, activities_to_zero) & (technosphere_matrix.row != technosphere_matrix.col)
 
-        for idx in row_idx:
-            # Skip the diagonal and exceptions
-            if tm_modified.row[idx] != act and (
-                exceptions is None or tm_modified.col[idx] not in exceptions
-            ):
-                tm_modified.data[idx] = 0
+    # Apply the mask to set the relevant elements to zero
+    print(technosphere_matrix.data[mask].sum() * -1)
+    technosphere_matrix.data[mask] = 0
 
-    tm_modified = tm_modified.tocsr()
-    tm_modified.eliminate_zeros()
+    mask_infra = np.isin(technosphere_matrix.row, infra) & (technosphere_matrix.row != technosphere_matrix.col)
+    print(technosphere_matrix.data[mask_infra].sum() * -1)
+    technosphere_matrix.data[mask_infra] = 0
 
-    # Remove double accounting
-    lca.technosphere_matrix = tm_modified
-    lca.lci(demand=demand)
-    return lca
+    # Convert the modified matrix back to CSR format and eliminate explicit zeros
+    #tm_modified = tm_modified.tocsr()
+    technosphere_matrix.eliminate_zeros()
 
+    # Update the LCA object's technosphere matrix and recalculate the inventory
+    #technosphere_matrix.technosphere_matrix = tm_modified
+    #technosphere_matrix.lci(demand=demand)
+    return technosphere_matrix.tocsr()
+
+# def remove_double_accounting(
+#     lca: bc.LCA,
+#     demand: Dict,
+#     activities_to_exclude: List[int],
+#     exceptions: List[int],
+# ):
+#     """
+#     Remove double counting from a technosphere matrix by zeroing out specified row values
+#     in all columns, except for those on the diagonal and those in the exceptions list.
+#
+#     :param lca: bw2calc.LCA object
+#     :param demand: dict with demand values
[…]
