+++ pathways/utils.py
@@ -480,6 +480,20 @@ def fetch_indices(
 
     return vars_idx
 
+def get_all_indices(vars_info: dict) -> list[int]:
+    """
+    Extract all 'idx' values from the vars_info dictionary.
+
+    :param vars_info: Dictionary of variables information returned by fetch_indices.
+    :type vars_info: dict
+    :return: List of all 'idx' values.
+    :rtype: list[int]
+    """
+    idx_list = []
+    for region_data in vars_info.values():
+        for variable_data in region_data.values():
+            idx_list.append(variable_data["idx"])
+    return idx_list
 
 def fetch_inventories_locations(
     technosphere_indices: Dict[str, Tuple[str, str, str]]
