+++ pathways/lca.py
@@ -103,7 +103,13 @@ def get_lca_matrices(
     model: str,
     scenario: str,
     year: int,
-) -> Tuple[Datapackage, Dict, Dict, list[tuple[int, int]]]:
+    mapping: Dict = None,
+    regions: List[str] = None,
+    variables: List[str] = None,
+    geo: Geomap = None,
+    remove_infrastructure: bool = False,
+    double_counting: dict = None,
+) -> tuple[Datapackage, dict[tuple[str, str, str, str], int], dict[tuple, int], list[tuple[int, int]], dict | None]:
     """
     Retrieve Life Cycle Assessment (LCA) matrices from disk.
 
@@ -115,6 +121,7 @@ def get_lca_matrices(
     :type scenario: str
     :param year: The year of the scenario.
     :type year: int
+    :param remove_infrastructure: Whether to remove infrastructure exchanges from the technosphere matrix.
     :rtype: Tuple[sparse.csr_matrix, sparse.csr_matrix, Dict, Dict, List]
     """
 
@@ -147,6 +154,12 @@ def get_lca_matrices(
     # remove the last element of the tuple, which is the index
     biosphere_inds = {k[:-1]: v for k, v in biosphere_inds.items()}
 
+    # Fetch indices
+    if geo is not None:
+        vars_info = fetch_indices(mapping, regions, variables, technosphere_inds, geo)
+    else:
+        vars_info = None
+
     dp = bwp.create_datapackage()
 
     fp_A = select_filepath("A_matrix", [fp for fp in fps if "index" not in fp.name])
@@ -155,6 +168,47 @@ def get_lca_matrices(
     # Load matrices and add them to the datapackage
     for matrix_name, fp in [("technosphere_matrix", fp_A), ("biosphere_matrix", fp_B)]:
         data, indices, sign, distributions = load_matrix_and_index(fp)
+
+        if remove_infrastructure is True and matrix_name == "technosphere_matrix":
+            print("--------- Removing infrastructure exchanges")
+            infrastructure_indices = np.array([j for i, j in technosphere_inds.items() if i[2] == "unit" and "uranium" not in i[0].lower()])
+            idx = np.where(np.isin(indices["row"], infrastructure_indices) & (indices["row"] != indices["col"]))
+            data[idx] = 0
+
+        if double_counting is not None and matrix_name == "technosphere_matrix":
+            print("--------- Correcting for double-counting")
+            # remove double counting
+            for k, v in double_counting.items():
+                for region, values in vars_info.items():
+                    if k in values:
+                        idx_k = values[k]["idx"]
+                        for activity, y in v.items():
+                            if activity in values:
+                                idx_v = values[activity]["idx"]
+                                amount = y[year]
+                                # add indices to `indices`
+                                indices = np.append(indices, np.array([(idx_k, idx_v)], dtype=bwp.INDICES_DTYPE), axis=0)
+                                data = np.append(data, amount)
+                                print(f"Added {amount} from {k} to {activity}")
+                                distributions=np.append(
+                                    distributions,
+                                    np.array(
+                                        (
+                                            0,
+                                            None,
+                                            None,
+                                            None,
+                                            None,
+                                            None,
+                                            False
+                                        ),
+                                        dtype=bwp.UNCERTAINTY_DTYPE,
+                                    )
+                                )
+                                sign = np.append(
+                                    sign,
[…]
