### 2024-06-17 cbc3419
Add `remove_uncertainty` argument to calculate()
