### 2024-06-17 9eca832
Update activities_classifications.yaml
