### 2024-07-03 dd08170
Update activities_classifications.yaml
