### 2024-07-03 959d859
Update activities_classifications.yaml
