+++ pathways/lca.py
@@ -23,7 +23,6 @@ from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, STATS_DIR, USER_LOGS_
 from .lcia import fill_characterization_factors_matrices
 from .stats import (
     create_mapping_sheet,
-    log_double_accounting,
     log_intensities_to_excel,
     log_results_to_excel,
     log_subshares_to_excel,
@@ -37,13 +36,9 @@ from .subshares import (
 from .utils import (
     CustomFilter,
     _group_technosphere_indices,
-    apply_filters,
     check_unclassified_activities,
     fetch_indices,
-    get_all_indices,
-    get_combined_filters,
     get_unit_conversion_factors,
-    read_categories_from_yaml,
     read_indices_csv,
 )
 
@@ -108,15 +103,13 @@ def get_lca_matrices(
     regions: List[str] = None,
     variables: List[str] = None,
     geo: Geomap = None,
-    remove_infrastructure: bool = False,
-    double_counting: dict = None,
     remove_uncertainty: bool = False,
 ) -> tuple[
     Datapackage,
     dict[tuple[str, str, str, str], int],
     dict[tuple, int],
     list[tuple[int, int]],
-    dict | None,
+    [dict, None],
 ]:
     """
     Retrieve Life Cycle Assessment (LCA) matrices from disk.
@@ -186,48 +179,7 @@ def get_lca_matrices(
                 dtype=bwp.UNCERTAINTY_DTYPE,
             )
 
-        if remove_infrastructure is True and matrix_name == "technosphere_matrix":
-            print("--------- Removing infrastructure exchanges")
-            infrastructure_indices = np.array(
-                [
-                    j
-                    for i, j in technosphere_inds.items()
-                    if i[2] == "unit" and "uranium" not in i[0].lower()
-                ]
-            )
-            idx = np.where(
-                np.isin(indices["row"], infrastructure_indices)
-                & (indices["row"] != indices["col"])
-            )
-            data[idx] = 0
-
-        if double_counting is not None and matrix_name == "technosphere_matrix":
-            print("--------- Correcting for double-counting")
-            # remove double counting
-            for k, v in double_counting.items():
-                for region, values in vars_info.items():
-                    if k in values:
-                        idx_k = values[k]["idx"]
-                        for activity, y in v.items():
-                            if activity in values:
-                                idx_v = values[activity]["idx"]
-                                amount = y[year]
-                                # add indices to `indices`
-                                indices = np.append(
-                                    indices,
-                                    np.array([(idx_k, idx_v)], dtype=bwp.INDICES_DTYPE),
-                                    axis=0,
-                                )
-                                data = np.append(data, amount)
-                                distributions = np.append(
-                                    distributions,
-                                    np.array(
-                                        (0, None, None, None, None, None, False),
-                                        dtype=bwp.UNCERTAINTY_DTYPE,
-                                    ),
-                                )
-                                sign = np.append(sign, np.array([True]))
-
+        uncertain_parameters = None
         if matrix_name == "technosphere_matrix":
             uncertain_parameters = find_uncertain_parameters(distributions, indices)
 
@@ -270,8 +222,8 @@ def remove_double_counting(
     :return: Technosphere matrix with double counting removed
     """
 
-    # Copy and convert the technosphere matrix to COO format for easy manipulation
-    # technosphere_matrix_original = technosphere_matrix.technosphere_matrix.copy()
+    # Copy and convert the technosphere matrix
[…]
