### 2024-07-05 e2666ae
Code cleanup
