+++ pathways/lca.py
@@ -25,8 +25,15 @@ from .stats import (
     log_intensities_to_excel,
     log_results_to_excel,
     log_subshares_to_excel,
-    run_stats_analysis,
+    run_GSA_OLS,
+    run_GSA_delta,
 )
+
+import time
+
+# from .montecarlo import MonteCarloLCA
+# from .sensitivity_analysis import GlobalSensitivityAnalysis
+
 from .subshares import (
     adjust_matrix_based_on_shares,
     find_technology_indices,
@@ -260,6 +267,7 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
         debug,
         use_distributions,
         uncertain_parameters,
+        # mc_lca,
     ) = data
 
     variables_demand = {}
@@ -268,6 +276,21 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
     param_keys = set()
     params_container = []
 
+    total_demand = scenarios.sel(
+        region=region,
+        model=model,
+        pathway=scenario,
+        year=year,
+    ).sum(dim="variables").values
+
+    if total_demand == 0:
+        print(
+            f"Total demand for {region}, {model}, {scenario}, {year} is zero. "
+            f"Skipping."
+        )
+        return d
+
+    variable_count = 0
     for v, variable in enumerate(variables):
         idx, dataset = vars_idx[variable]["idx"], vars_idx[variable]["dataset"]
         # Compute the unit conversion vector for the given activities
@@ -278,7 +301,7 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
             scenarios.attrs["units"][variable],
             dataset_unit,
             units_map,
-        )
+        ).astype(float)
 
         # Fetch the demand for the given
         # region, model, pathway, and year
@@ -288,18 +311,11 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
             model=model,
             pathway=scenario,
             year=year,
-        )
+        ).values * unit_vector
 
-        total_demand = scenarios.sel(
-            region=region,
-            model=model,
-            pathway=scenario,
-            year=year,
-        ).sum(dim="variables")
-
-        if total_demand == 0:
+        if np.sum(demand) == 0:
             print(
-                f"Total demand for {region}, {model}, {scenario}, {year} is zero. "
+                f"Total demand for {variable} in {region}, {model}, {scenario}, {year} is zero. "
                 f"Skipping."
             )
             continue
@@ -313,10 +329,10 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
 
         variables_demand[variable] = {
             "id": idx,
-            "demand": demand.values * float(unit_vector),
+            "demand": demand,
         }
 
-        functional_unit = {idx: demand.values * float(unit_vector)}
+        functional_unit = {idx: demand}
 
         with CustomFilter("(almost) singular matrix"):
             lca.lci(demand=functional_unit)
@@ -326,6 +342,13 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
             characterized_inventory = (
                 characterization_matrix @ lca.inventory
             ).toarray()
+        # else:
+        #     mc_lca.func_units = functional_unit
+        #     mc_lca.methods = methods
+        #     mc_lca.calculate(iterations=use_distributions, seed=42)
+        #     results = mc_lca.results
+        #     characterized_inventory = np.quantile(results, [0.05, 0.5, 0.95], axis=0)
+        #     mc_lca.store_results()  # Save the Monte Carlo results for GSA
 
         else:
[…]
