### 2024-07-22 bfce1bf
Code cleanup
