+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy<2.0.0
+numpy==1.24.0
 pathlib
 pandas
 scipy
@@ -7,8 +7,10 @@ premise
 pyyaml
 bw_processing
 bw2calc >= 2.0.dev17
+scikit-umfpack
 datapackage
 pyprind
 platformdirs
 fs
 statsmodels
+SALib
