+++ conda/meta.yaml
@@ -31,6 +31,7 @@ requirements:
     - pyprind
     - platformdirs
     - fs
+    - SALib
 test:
   imports:
     - pathways
