+++ dev/timing.py
@@ -5,9 +5,11 @@ p = Pathways(datapackage="remind-SSP2-PkBudg1150-stem-SPS1.zip")
 
 p.calculate(
     methods=[
-        'EF v3.1 EN15804 - climate change - global warming potential (GWP100)',
+        "EF v3.1 EN15804 - climate change - global warming potential (GWP100)",
     ],
-    regions=["CH",],
+    regions=[
+        "CH",
+    ],
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[
         2050,
@@ -15,5 +17,5 @@ p.calculate(
     variables=[v for v in p.scenarios.coords["variables"].values if v.startswith("FE")],
     use_distributions=20,
     subshares=True,
-    multiprocessing=False
+    multiprocessing=False,
 )
