### 2024-07-22 2fcbac7
Code cleanup
