+++ pathways/lca.py
@@ -7,6 +7,7 @@ import logging
 import uuid
 from pathlib import Path
 from typing import Any, Dict, List, Tuple
+from collections import defaultdict
 
 import bw2calc as bc
 import bw_processing as bwp
@@ -270,9 +271,7 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
 
     variables_demand = {}
     d = []
-    impacts_by_method = {method: [] for method in methods}
-    param_keys = set()
-    params_container = []
+
 
     total_demand = (
         scenarios.sel(
@@ -292,7 +291,6 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
         )
         return d
 
-    variable_count = 0
     for v, variable in enumerate(variables):
         idx, dataset = vars_idx[variable]["idx"], vars_idx[variable]["dataset"]
         # Compute the unit conversion vector for the given activities
@@ -335,70 +333,74 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
         variables_demand[variable] = {
             "id": idx,
             "demand": demand,
+            "fu": {idx: demand},
+            "dataset": dataset,
+            "unit vector": unit_vector,
         }
 
-        functional_unit = {idx: demand}
+    param_keys = set()
 
+
+    if use_distributions == 0:
+        # Regular LCA calculations
         with CustomFilter("(almost) singular matrix"):
-            lca.lci(demand=functional_unit)
+            for key, value in variables_demand.items():
+                lca.lci(demand=value["fu"])
 
-        if use_distributions == 0:
-            # Regular LCA
-            characterized_inventory = (
-                characterization_matrix @ lca.inventory
-            ).toarray()
+                characterized_inventory = (
+                    characterization_matrix @ lca.inventory
+                ).toarray()
+                d.append(characterized_inventory)
 
-        else:
-            # Use distributions for LCA calculations
-            # next(lca) is a generator that yields the inventory matrix
-            temp_results = []
-            params = {}
-
-            for _ in zip(range(use_distributions), lca):
-
-                matrix_result = (characterization_matrix @ lca.inventory).toarray()
-                temp_results.append(matrix_result)
-                for i in range(len(uncertain_parameters)):
-                    param_key = (
-                        f"{uncertain_parameters[i][0]}_to_{uncertain_parameters[i][1]}"
+                if debug:
+                    logging.info(
+                        f"var.: {key}, name: {value['dataset'][0][:50]}, "
+                        f"ref.: {value['dataset'][1]}, unit: {value['dataset'][2][:50]}, idx: {value['idx']},"
+                        f"loc.: {value['dataset'][3]}, demand: {value['demand']}, "
+                        f"unit conv.: {value['unit vector']}, "
+                        f"impact: {np.round(characterized_inventory.sum(axis=-1) / value['demand'], 3)}. "
                     )
-                    param_keys.add(param_key)
-                    if param_key not in params:
-                        params[param_key] = []
-                    value = -lca.technosphere_matrix[
-                        uncertain_parameters[i][0], uncertain_parameters[i][1]
-                    ]
-                    params[param_key].append(value)
 
-            params_container.append(params)
+    else:
+        # Use distributions for LCA calculations
+        # next(lca) is a generator that yields the inventory matrix
+        results = np.zeros(
+            (use_distributions, len(variables_demand), len(methods))
+        )
+        params = defaultdict(list)
 
-            results = np.array(temp_results)
-            for idx, method in enumerate(methods):
-                # Sum over items for each stochastic result, not across all results
[…]
