### 2024-07-23 fb4560f
Code cleanup
