+++ pathways/stats.py
@@ -97,7 +97,7 @@ def log_double_accounting(
             )
 
 
-def log_subshares_to_excel(year: int, shares: dict, export_path: Path):
+def log_subshares_to_excel(year: int, shares: dict, total_impacts_df: pd.DataFrame) -> pd.DataFrame:
     """
     Logs results to an Excel file named according to model, scenario, and year, specifically for the given year.
     This method assumes that each entry in the shares defaultdict is structured to be directly usable in a DataFrame.
@@ -105,7 +105,7 @@ def log_subshares_to_excel(year: int, shares: dict, export_path: Path):
     Parameters:
     :param year: The specific year for which data is being logged.
     :param shares: A nested defaultdict containing shares data for multiple years and types.
-    :param export_path: The path to the Excel file where the data will be logged.
+    :param total_impacts_df: DataFrame containing total impacts for each method.
     """
 
     data = []
@@ -128,32 +128,24 @@ def log_subshares_to_excel(year: int, shares: dict, export_path: Path):
 
     new_df = pd.DataFrame(data)
 
-    try:
-        if os.path.exists(export_path):
-            df_existing = pd.read_excel(export_path)
-            # Merge new data into existing data, selectively updating share columns
-            combined_df = (
-                df_existing.set_index(["Iteration", "Year"])
-                .combine_first(new_df.set_index(["Iteration", "Year"]))
-                .reset_index()
-            )
-            # Optionally, ensure the columns are in a meaningful order
-            new_columns = [
-                col for col in new_df.columns if col not in ["Iteration", "Year"]
-            ]
-            existing_columns = [
-                col for col in df_existing.columns if col not in new_df.columns
-            ]
-            combined_df = combined_df[
-                ["Iteration", "Year"] + new_columns + existing_columns
-            ]
+    # Merge new data into existing data, selectively updating share columns
+    combined_df = (
+        total_impacts_df.set_index(["Iteration", "Year"])
+        .combine_first(new_df.set_index(["Iteration", "Year"]))
+        .reset_index()
+    )
+    # Optionally, ensure the columns are in a meaningful order
+    new_columns = [
+        col for col in new_df.columns if col not in ["Iteration", "Year"]
+    ]
+    existing_columns = [
+        col for col in total_impacts_df.columns if col not in new_df.columns
+    ]
+    combined_df = combined_df[
+        ["Iteration", "Year"] + new_columns + existing_columns
+    ]
 
-            combined_df.to_excel(export_path, index=False)
-        else:
-            new_df.to_excel(export_path, index=False)
-    except Exception as e:
-        print(f"Error updating Excel file: {str(e)}")
-
+    return combined_df
 
 def log_intensities_to_excel(year: int, params: list, export_path: Path):
     """
@@ -207,7 +199,6 @@ def log_intensities_to_excel(year: int, params: list, export_path: Path):
 def log_results_to_excel(
     total_impacts_by_method: dict,
     methods: list,
-    filepath: Path,
 ):
     """
     Log the characterized inventory results for each LCIA method into separate columns in an Excel file.
@@ -218,10 +209,8 @@ def log_results_to_excel(
     :param filepath: Optional. File path for the Excel file to save the results.
     """
 
-    try:
-        df = pd.read_excel(filepath)
-    except FileNotFoundError:
-        df = pd.DataFrame()
+
+    df = pd.DataFrame()
 
     for method, impacts in total_impacts_by_method.items():
         df[method] = pd.Series(impacts)
@@ -230,7 +219,10 @@ def log_results_to_excel(
[…]
