+++ pathways/pathways.py
@@ -19,6 +19,12 @@ from .data_validation import validate_datapackage
 from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, STATS_DIR, USER_LOGS_DIR
 from .lca import _calculate_year, get_lca_matrices
 from .lcia import get_lcia_method_names
+from .stats import (
+    create_mapping_sheet,
+    log_results_to_excel,
+    log_subshares_to_excel,
+    run_GSA_delta,
+)
 from .subshares import generate_samples
 from .utils import (
     _get_mapping,
@@ -34,8 +40,6 @@ from .utils import (
     load_units_conversion,
     resize_scenario_data,
 )
-
-from .stats import log_results_to_excel, log_subshares_to_excel, create_mapping_sheet, run_GSA_delta
 
 
 class Pathways:
@@ -381,7 +385,9 @@ class Pathways:
 
         self._fill_in_result_array(results, use_distributions, subshares, methods)
 
-    def _fill_in_result_array(self, results: dict, use_distributions: int, subshares: bool, methods: list) -> None:
+    def _fill_in_result_array(
+        self, results: dict, use_distributions: int, subshares: bool, methods: list
+    ) -> None:
 
         # Assuming DIR_CACHED_DB, results, and self.lca_results are already defined
 
@@ -460,7 +466,9 @@ class Pathways:
                     # create new worksheet called 'Total impacts'
                     writer.book.create_sheet("Total impacts")
                     # add df to it
-                    df_tot_impacts.to_excel(writer, sheet_name="Total impacts", index=True)
+                    df_tot_impacts.to_excel(
+                        writer, sheet_name="Total impacts", index=True
+                    )
 
                     if subshares:
                         df_shares = log_subshares_to_excel(
