+++ pathways/lca.py
@@ -18,6 +18,7 @@ from bw_processing import Datapackage
 from numpy import dtype, ndarray
 from premise.geomap import Geomap
 from scipy import sparse
+from bw2calc import MultiLCA
 
 from .filesystem_constants import DIR_CACHED_DB, STATS_DIR, USER_LOGS_DIR
 from .lcia import fill_characterization_factors_matrices
@@ -243,35 +244,19 @@ def remove_double_counting(
     return technosphere_matrix.tocsr()
 
 
-def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int]]:
-    """
-    Process the region data.
-    :param data: Tuple containing the model, scenario, year, region, variables, vars_idx, scenarios, units_map,
-                    demand_cutoff, lca, characterization_matrix, debug, use_distributions, uncertain_parameters.
-    :return: Dictionary containing the region data.
-    """
-    (
+def create_functional_units(
+        scenarios,
+        region,
         model,
         scenario,
         year,
-        region,
         variables,
         vars_idx,
-        scenarios,
         units_map,
-        demand_cutoff,
-        lca,
-        characterization_matrix,
-        methods,
-        debug,
-        use_distributions,
-        uncertain_parameters,
-        # mc_lca,
-    ) = data
+        demand_cutoff
+) -> [dict, dict]:
 
     variables_demand = {}
-    d = []
-
 
     total_demand = (
         scenarios.sel(
@@ -289,7 +274,6 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
             f"Total demand for {region}, {model}, {scenario}, {year} is zero. "
             f"Skipping."
         )
-        return d
 
     for v, variable in enumerate(variables):
         idx, dataset = vars_idx[variable]["idx"], vars_idx[variable]["dataset"]
@@ -306,14 +290,14 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
         # Fetch the demand for the given
         # region, model, pathway, and year
         demand = (
-            scenarios.sel(
-                variables=variable,
-                region=region,
-                model=model,
-                pathway=scenario,
-                year=year,
-            ).values
-            * unit_vector
+                scenarios.sel(
+                    variables=variable,
+                    region=region,
+                    model=model,
+                    pathway=scenario,
+                    year=year,
+                ).values
+                * unit_vector
         )
 
         if np.sum(demand) == 0:
@@ -338,69 +322,94 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
             "unit vector": unit_vector,
         }
 
-    param_keys = set()
+    return {
+        key: value["fu"] for key, value in variables_demand.items()
+    }, variables_demand
 
+
+def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int]]:
+    """
+    Process the region data.
+    :param data: Tuple containing the model, scenario, year, region, variables, vars_idx, scenarios, units_map,
+                    demand_cutoff, lca, characterization_matrix, debug, use_distributions, uncertain_parameters.
+    :return: Dictionary containing the region data.
+    """
+    (
+        model,
+        scenario,
+        year,
+        region,
+        variables,
+        fus_details,
+        scenarios,
+        units_map,
+        demand_cutoff,
+        lca,
+        characterization_matrix,
+        methods,
+        debug,
+        use_distributions,
+        uncertain_parameters,
+    ) = data
+
+    d = []
 
     if use_distributions == 0:
         # Regular LCA calculations
         with CustomFilter("(almost) singular matrix"):
-            for key, value in variables_demand.items():
[…]

[…]
