+++ pathways/stats.py
@@ -97,7 +97,9 @@ def log_double_accounting(
             )
 
 
-def log_subshares_to_excel(year: int, shares: dict, total_impacts_df: pd.DataFrame) -> pd.DataFrame:
+def log_subshares_to_excel(
+    year: int, shares: dict, total_impacts_df: pd.DataFrame
+) -> pd.DataFrame:
     """
     Logs results to an Excel file named according to model, scenario, and year, specifically for the given year.
     This method assumes that each entry in the shares defaultdict is structured to be directly usable in a DataFrame.
@@ -135,19 +137,18 @@ def log_subshares_to_excel(year: int, shares: dict, total_impacts_df: pd.DataFra
         .reset_index()
     )
     # Optionally, ensure the columns are in a meaningful order
-    new_columns = [
-        col for col in new_df.columns if col not in ["Iteration", "Year"]
-    ]
+    new_columns = [col for col in new_df.columns if col not in ["Iteration", "Year"]]
     existing_columns = [
         col for col in total_impacts_df.columns if col not in new_df.columns
     ]
-    combined_df = combined_df[
-        ["Iteration", "Year"] + new_columns + existing_columns
-    ]
+    combined_df = combined_df[["Iteration", "Year"] + new_columns + existing_columns]
 
     return combined_df
 
-def log_intensities_to_excel(year: int, params: list, df_tot_impacts: pd.DataFrame) -> pd.DataFrame:
+
+def log_intensities_to_excel(
+    year: int, params: list, df_tot_impacts: pd.DataFrame
+) -> pd.DataFrame:
     """
     Update or create an Excel file with new columns of data, based on model, scenario, and year.
 
@@ -160,7 +161,6 @@ def log_intensities_to_excel(year: int, params: list, df_tot_impacts: pd.DataFra
         print("Warning: No new data provided to log.")
         return
 
-
     # merge list of dictionaries into a single dictionary
     params = {k: v for d in params for k, v in d.items()}
 
@@ -169,7 +169,6 @@ def log_intensities_to_excel(year: int, params: list, df_tot_impacts: pd.DataFra
     df_new = pd.DataFrame(params)
     df_new["Iteration"] = range(1, max_length + 1)
     df_new["Year"] = [year] * max_length
-
 
     combined_df = pd.merge(
         df_tot_impacts,
@@ -206,18 +205,16 @@ def log_results_to_excel(
     df = pd.DataFrame(columns=methods + ["Iteration", "Year", "Region"])
 
     for (region, year), values in total_impacts_by_method.items():
-        df = pd.concat([
-            df,
-            pd.DataFrame(
-                values, columns=methods
-            ).assign(Iteration=range(1, len(values) + 1), Year=year, Region=region)
+        df = pd.concat(
+            [
+                df,
+                pd.DataFrame(values, columns=methods).assign(
+                    Iteration=range(1, len(values) + 1), Year=year, Region=region
+                ),
             ]
         )
 
     return df
-
-
-
 
 
 def create_mapping_sheet(
@@ -368,7 +365,10 @@ def run_GSA_delta(methods: list, df_tot_impacts: pd.DataFrame) -> pd.DataFrame:
     problem = {
         "num_vars": len(params),
         "names": params,
-        "bounds": [[df_tot_impacts[param].min(), df_tot_impacts[param].max()] for param in params],
+        "bounds": [
+            [df_tot_impacts[param].min(), df_tot_impacts[param].max()]
+            for param in params
+        ],
     }
 
     results = []
@@ -381,8 +381,6 @@ def run_GSA_delta(methods: list, df_tot_impacts: pd.DataFrame) -> pd.DataFrame:
         Y = df_tot_impacts[method].values
 
[…]
