### 2024-07-24 682923f
Code cleanup
