+++ pathways/stats.py
@@ -97,176 +97,86 @@ def log_double_accounting(
             )
 
 
-def log_subshares_to_excel(
-    year: int, shares: dict, total_impacts_df: pd.DataFrame
+def log_subshares(
+    shares: dict, region: str,
 ) -> pd.DataFrame:
     """
-    Logs results to an Excel file named according to model, scenario, and year, specifically for the given year.
-    This method assumes that each entry in the shares defaultdict is structured to be directly usable in a DataFrame.
-
-    Parameters:
-    :param year: The specific year for which data is being logged.
-    :param shares: A nested defaultdict containing shares data for multiple years and types.
-    :param total_impacts_df: DataFrame containing total impacts for each method.
+    Create a pandas DataFrame where the keys of shares are the columns
+    and the values are the values. The region is added as a column.
     """
 
-    data = []
+    df = pd.DataFrame(shares)
+    df["region"] = region
+    df["iteration"] = range(1, len(df) + 1)
 
-    first_tech = next(iter(shares), None)
-    if not first_tech or year not in shares[first_tech]:
-        print(f"No data found for year {year} in any technology.")
-        return
-
-    num_iterations = len(shares[first_tech][year][next(iter(shares[first_tech][year]))])
-    for i in range(num_iterations):
-        iteration_data = {"Iteration": i + 1, "Year": year}
-        for tech, years_data in shares.items():
-            if year in years_data:
-                for subtype, values in years_data[year].items():
-                    iteration_data[f"{tech}_{subtype}"] = (
-                        values[i] if i < len(values) else None
-                    )
-        data.append(iteration_data)
-
-    new_df = pd.DataFrame(data)
-
-    # Merge new data into existing data, selectively updating share columns
-    combined_df = (
-        total_impacts_df.set_index(["Iteration", "Year"])
-        .combine_first(new_df.set_index(["Iteration", "Year"]))
-        .reset_index()
-    )
-    # Optionally, ensure the columns are in a meaningful order
-    new_columns = [col for col in new_df.columns if col not in ["Iteration", "Year"]]
-    existing_columns = [
-        col for col in total_impacts_df.columns if col not in new_df.columns
-    ]
-    combined_df = combined_df[["Iteration", "Year"] + new_columns + existing_columns]
-
-    return combined_df
+    return df[["iteration", "region"] + list(shares.keys())]
 
 
-def log_intensities_to_excel(
-    year: int, params: list, df_tot_impacts: pd.DataFrame
+def log_uncertainty_values(
+        region: str,
+        uncertainty_indices: np.array,
+        uncertainty_values: np.array,
 ) -> pd.DataFrame:
     """
-    Update or create an Excel file with new columns of data, based on model, scenario, and year.
+    Create a pandas DataFrame with the region and uncertainty indices as columns,
+    the uncertainty values as values, and the iteration number as the index.
 
-    :param year: The year for which the data is logged.
-    :param params: Dictionary where keys are the new column names and values are lists of data for each column.
+    :param region: Name of the region
+    :param uncertainty_indices: Indices of the uncertainty values
+    :param uncertainty_values: Uncertainty values
+    :return: DataFrame with region as column, uncertainty indices as indices and uncertainty values as values
 
     """
 
-    if not params:
-        print("Warning: No new data provided to log.")
-        return
+    # convert 2D numpy array to list of tuples
+    uncertainty_indices = uncertainty_indices.tolist()
+    uncertainty_indices = [[str(x) for x in index] for index in uncertainty_indices]
[…]
