+++ dev/timing.py
@@ -15,7 +15,7 @@ p.calculate(
     ],
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[
-        2050,
+        2035,
     ],
     variables=vars,
     use_distributions=3,
