+++ pathways/pathways.py
@@ -14,6 +14,7 @@ import pandas as pd
 import pyprind
 import xarray as xr
 import yaml
+import pickle
 
 from .data_validation import validate_datapackage
 from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, STATS_DIR, USER_LOGS_DIR
@@ -21,9 +22,9 @@ from .lca import _calculate_year, get_lca_matrices
 from .lcia import get_lcia_method_names
 from .stats import (
     create_mapping_sheet,
-    log_intensities_to_excel,
-    log_results_to_excel,
-    log_subshares_to_excel,
+    log_uncertainty_values,
+    log_results,
+    log_subshares,
     run_GSA_delta,
 )
 from .subshares import generate_samples
@@ -384,10 +385,10 @@ class Pathways:
         # remove None values in results
         results = {k: v for k, v in results.items() if v is not None}
 
-        self._fill_in_result_array(results, use_distributions, subshares, methods)
+        self._fill_in_result_array(results, use_distributions, shares, methods)
 
     def _fill_in_result_array(
-        self, results: dict, use_distributions: int, subshares: bool, methods: list
+        self, results: dict, use_distributions: int, shares: [None, dict], methods: list
     ) -> None:
 
         # Assuming DIR_CACHED_DB, results, and self.lca_results are already defined
@@ -409,19 +410,15 @@ class Pathways:
             model, scenario, year = coord
             acts_category_idx_dict = result["other"]["acts_category_idx_dict"]
             acts_location_idx_dict = result["other"]["acts_location_idx_dict"]
-            total_impacts = {}
 
-            for region, data in result.items():
+            for region, data_ in result.items():
                 if region == "other":
                     continue
 
-                id_array = data["id_array"]
-                variables = data["variables"]
+                id_array = data_["id_array"]
+                variables = data_["variables"]
 
                 d = np.squeeze(cached_data[id_array])
-
-                if use_distributions > 0:
-                    d = np.quantile(d, [0.05, 0.5, 0.95], axis=0)
 
                 for cat, act_cat_idx in acts_category_idx_dict.items():
                     for loc, act_loc_idx in acts_location_idx_dict.items():
@@ -429,12 +426,6 @@ class Pathways:
                         idx = idx[idx != -1]
 
                         if idx.size > 0:
-
-                            if use_distributions > 0:
-                                total_impacts[(region, year)] = d.sum(axis=-1).sum(
-                                    axis=1
-                                )
-
                             summed_data = d[..., idx].sum(axis=-1)
 
                             try:
@@ -464,71 +455,120 @@ class Pathways:
                                     }
                                 ] = summed_data.transpose(1, 2, 0)
 
-            if use_distributions > 0:
+        if use_distributions > 0:
+
+            uncertainty_parameters = {
+                data["uncertainty_params"]: load_numpy_array_from_disk(
+                    DIR_CACHED_DB / f"{data['uncertainty_params']}.npy",
+                )
+                for coord, result in results.items()
+                for region, data in result.items()
+                if region != "other"
+            }
+
+            uncertainty_values = {
+                data["uncertainty_vals"]: load_numpy_array_from_disk(
+                    DIR_CACHED_DB / f"{data['uncertainty_vals']}.npy",
+                )
+                for coord, result in results.items()
+                for region, data in result.items()
+                if region != "other"
+            }
+
+            tehnosphere_indices = {
+                data["technosphere_indices"]: pickle.load(
+                    open(DIR_CACHED_DB / f"{data['technosphere_indices']}.pkl", "rb")
+                )
+                for coord, result in results.items()
+                for region, data in result.items()
+                if region != "other"
+            }
+
[…]

[…]
