+++ pathways/lcia.py
@@ -11,7 +11,7 @@ from scipy.sparse import csr_matrix
 
 from .filesystem_constants import DATA_DIR
 
-LCIA_METHODS = DATA_DIR / "lcia_data.json"
+LCIA_METHODS = DATA_DIR / "lcia_ei310.json"
 
 
 def get_lcia_method_names():
