### 2024-07-25 99452f8
Add scrip to update LCIA methods
