### 2024-07-27 00c69bf
Add scrip to update LCIA methods
