+++ pathways/lca.py
@@ -20,6 +20,7 @@ from bw_processing import Datapackage
 from numpy import dtype, ndarray
 from premise.geomap import Geomap
 from scipy import sparse
+import sparse as sp
 
 from .filesystem_constants import DIR_CACHED_DB, STATS_DIR, USER_LOGS_DIR
 from .lcia import fill_characterization_factors_matrices
@@ -300,20 +301,6 @@ def create_functional_units(
             * unit_vector
         )
 
-        if np.sum(demand) == 0:
-            logging.info(
-                f"Total demand for {variable} in {region}, {model}, {scenario}, {year} is zero. "
-                f"Skipping."
-            )
-            continue
-
-        # If the demand is below the cut-off criteria, skip to the next iteration
-        share = demand / total_demand
-
-        # If the total demand is zero, return None
-        if share < demand_cutoff:
-            continue
-
         variables_demand[variable] = {
             "id": idx,
             "demand": demand,
@@ -327,7 +314,7 @@ def create_functional_units(
     }, variables_demand
 
 
-def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int]]:
+def process_region(data: Tuple) -> Dict[str, str | List[str] | List[int]]:
     """
     Process the region data.
     :param data: Tuple containing the model, scenario, year, region, variables, vars_idx, scenarios, units_map,
@@ -352,94 +339,121 @@ def process_region(data: Tuple) -> dict[str, ndarray[Any, dtype[Any]] | list[int
         uncertain_parameters,
     ) = data
 
-    lci_results = []
     id_uncertainty_indices = None
-    id_uncertainty_values = None
     id_technosphere_indices = None
-    id_iter_results_array = None
+    iter_results_files = []
+    iter_param_vals_files = []
+
+    dict_loc_cat = {}
+
+    cat_counter = 0
+    for cat, act_cat_idx in lca.acts_category_idx_dict.items():
+        loc_counter = 0
+        for loc, act_loc_idx in lca.acts_location_idx_dict.items():
+            # Find the intersection of indices
+            idx = np.intersect1d(act_cat_idx, act_loc_idx)
+            # Filter out any -1 indices
+            filtered_idx = idx[idx != -1]
+
+            if filtered_idx.size > 0:
+                # Assign the filtered index array to the dict_loc_cat with (cat, loc) as key
+                dict_loc_cat[(cat_counter, loc_counter)] = filtered_idx
+
+            loc_counter += 1
+        cat_counter += 1
 
     if use_distributions == 0:
         # Regular LCA calculations
         with CustomFilter("(almost) singular matrix"):
             lca.lci()
 
-        characterized_inventories = {
-            key: (characterization_matrix @ value).toarray()
-            for key, value in lca.inventories.items()
-        }
+        # Create a numpy array with the results
+        inventory_results = np.array([
+            (characterization_matrix @ value).toarray()
+            for value in lca.inventories.values()
+        ])
 
-        lci_results.extend(list(characterized_inventories.values()))
+        iter_results = np.zeros(
+            (
+                inventory_results.shape[0],
+                inventory_results.shape[1],
+                len(lca.acts_category_idx_dict),
+                len(lca.acts_location_idx_dict),
+            )
+        )
+
+        for (cat, loc), idx in dict_loc_cat.items():
+            iter_results[:, :, cat, loc] = inventory_results[:, :, idx].sum(axis=2)
+
+        # Save iteration results to disk
+        iter_results_file = f"iter_results_{uuid.uuid4()}.npz"
+        sp.save_npz(filename=DIR_CACHED_DB / iter_results_file, matrix=sp.COO(iter_results), compressed=True)
+        iter_results_files.append(iter_results_file)
 
     else:
[…]
