+++ pathways/pathways.py
@@ -9,12 +9,13 @@ import pickle
 from collections import defaultdict
 from multiprocessing import Pool, cpu_count
 from typing import Any, List, Optional
+from pathlib import Path
 
 import numpy as np
 import pandas as pd
-import pyprind
 import xarray as xr
 import yaml
+import sparse as sp
 
 from .data_validation import validate_datapackage
 from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, STATS_DIR, USER_LOGS_DIR
@@ -38,10 +39,195 @@ from .utils import (
     fetch_inventories_locations,
     harmonize_units,
     load_classifications,
-    load_numpy_array_from_disk,
     load_units_conversion,
     resize_scenario_data,
 )
+
+
+def _fill_in_result_array(
+        coords: tuple, result: dict, use_distributions: int, shares: [None, dict], methods: list
+) -> np.ndarray:
+    def _load_array(filepath):
+        if len(filepath) == 1:
+            if Path(filepath[0]).suffix == ".npy":
+                return np.load(DIR_CACHED_DB / filepath[0])
+            elif Path(filepath[0]).suffix == ".pkl":
+                with open(DIR_CACHED_DB / filepath[0], "rb") as f:
+                    return pickle.load(f)
+            else:
+                return sp.load_npz(DIR_CACHED_DB / filepath[0]).todense()
+        else:
+            if Path(filepath[0]).suffix == ".npy":
+                return np.stack([
+                    np.load(DIR_CACHED_DB / f)
+                    for f in filepath
+                ], axis=-1)
+            return np.stack([
+                sp.load_npz(DIR_CACHED_DB / f).todense()
+                for f in filepath
+            ], axis=-1)
+
+    # Pre-loading data from disk if possible
+    cached_data = {
+        region: _load_array(data["iterations_results"])
+        for region, data in result.items()
+    }
+
+    uncertainty_parameters, uncertainty_values, tehnosphere_indices, iteration_results = None, None, None, None
+
+    if use_distributions > 0:
+        uncertainty_parameters = {
+            region: _load_array(data["uncertainty_params"])
+            for region, data in result.items()
+        }
+
+        uncertainty_values = {
+            region: _load_array(data["iterations_param_vals"])
+            for region, data in result.items()
+        }
+
+        tehnosphere_indices = {
+            region: _load_array(data["technosphere_indices"])
+            for region, data in result.items()
+        }
+
+        iteration_results = {
+            region: _load_array(data["iterations_results"])
+            for region, data in result.items()
+        }
+
+    model, scenario, year = coords
+
+    results = []
+
+    for region, data_ in result.items():
+
+        d = cached_data[region]
+
+        if use_distributions > 0:
+            d = np.quantile(d, [0.05, 0.5, 0.95], axis=-1)
+            d = d.transpose(3, 1, 4, 2, 0)
+        else:
+            d = d.transpose(2, 0, 3, 1)
+
+        results.append(d)
+
+    if use_distributions > 0:
+        statistical_analysis(
+            model=model,
+            scenario=scenario,
+            year=year,
+            methods=methods,
+            result=result,
+            uncertainty_parameters=uncertainty_parameters,
+            uncertainty_values=uncertainty_values,
+            tehnosphere_indices=tehnosphere_indices,
+            iteration_results=iteration_results,
+            shares=shares,
+        )
+
+    return np.stack(results, axis=2)
+
+
+def statistical_analysis(
+    model: str,
+    scenario: str,
+    year: int,
+    methods: list,
+    result: dict,
+    uncertainty_parameters: dict,
+    uncertainty_values: dict,
[…]
