+++ dev/timing.py
@@ -14,12 +14,7 @@ p.calculate(
         "CH",
     ],
     scenarios=p.scenarios.pathway.values.tolist(),
-    years=[
-        2020,
-        2030,
-        2040,
-        2050
-    ],
+    years=[2020, 2030, 2040, 2050],
     variables=vars,
     use_distributions=100,
     subshares=True,
