+++ dev/timing.py
@@ -15,9 +15,9 @@ p.calculate(
     ],
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[
-        #2020,
-        #2030,
-        #2040,
+        # 2020,
+        # 2030,
+        # 2040,
         2050
     ],
     variables=vars,
