+++ pathways/data/activities_classifications.yaml
@@ -76526,193 +76526,3 @@
 - carbon dioxide, captured at cement plant
 - kilogram
 : Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using monoethanolamine
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
-? !!python/tuple
-- carbon dioxide, captured at cement production plant, using oxyfuel
-- carbon dioxide, captured at cement plant
-- kilogram
-: Carbon capture
[…]
