+++ dev/timing.py
@@ -1,6 +1,9 @@
 from pathways import Pathways
 
-p = Pathways(datapackage="remind-SSP2-PkBudg1150-stem-SPS1.zip")
+p = Pathways(
+    datapackage="remind-SSP2-PkBudg1150-stem-SPS1.zip",
+    geography_mapping="geo_mapping_remind.yaml"
+)
 
 vars = [v for v in p.scenarios.coords["variables"].values if v.startswith("FE")]
 
@@ -15,15 +18,16 @@ p.calculate(
     ],
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[
-        # 2020,
-        # 2030,
-        # 2040,
+        2020,
+        2030,
+        2040,
         2050
     ],
     variables=vars,
-    use_distributions=0,
+    use_distributions=100,
     subshares=True,
     multiprocessing=True,
+    statistical_analysis=True,
 )
 
 p.export_results()
