### 2024-07-29 aab8bd8
GSA analysis is separated from `calculate()`. Now `run_gsa()` is called afterwards instead.
