+++ pathways/data/technologies_shares.yaml
@@ -1,32 +1,52 @@
-### When different technology configurations or alternatives are found,
-### we are taking the best performing technology in terms of GWP.
 Wind: ### Check if IAMs make the distinction between onshore and offshore
-  onshore:
-    name: electricity production, wind, <1MW turbine, onshore
+  onshore-DDPM:
+    name: electricity production, wind, 1-3MW turbine, onshore, direct drive
     reference product: electricity, high voltage
     unit: kilowatt hour
     share:
       2020:
-        loc: 0.96
+        loc: 0.19
       2050:
-        minimum: 0.83
-        maximum: 0.96
+        minimum: 0.17
+        maximum: 0.19
         uncertainty_type: uniform
-  offshore:
+  onshore-Gearbox:
+    name: electricity production, wind, 1-3MW turbine, onshore
+    reference product: electricity, high voltage
+    unit: kilowatt hour
+    share:
+      2020:
+        loc: 0.77
+      2050:
+        minimum: 0.67
+        maximum: 0.77
+        uncertainty_type: uniform
+  offshore-DDPM:
+    name: electricity production, wind, 1-3MW turbine, offshore, direct drive
+    reference product: electricity, high voltage
+    unit: kilowatt hour
+    share:
+      2020:
+        loc: 0.022
+      2050:
+        minimum: 0.022
+        maximum: 0.15
+        uncertainty_type: uniform
+  offshore-Gearbox:
     name: electricity production, wind, 1-3MW turbine, offshore
     reference product: electricity, high voltage
     unit: kilowatt hour
     share:
       2020:
-        loc: 0.04
+        loc: 0.018
       2050:
-        minimum: 0.04
-        maximum: 0.17
+        minimum: 0.018
+        maximum: 0.10
         uncertainty_type: uniform
 PV:
   c-Si:
-    name: electricity production, photovoltaic, at 93 kWp slanted-roof, multi-Si, laminated, integrated
-    reference product: electricity production, photovoltaic, at 93 kWp slanted-roof, multi-Si, laminated, integrated
+    name: electricity production, photovoltaic, photovoltaic slanted-roof installation, 3 kWp, multi-Si, laminated, integrated
+    reference product: electricity, low voltage
     unit: kilowatt hour
     share:
       2020:
@@ -36,8 +56,8 @@ PV:
         maximum: 0.9440
         uncertainty_type: uniform
   CdTe:
-    name: electricity production, photovoltaic, at 3kWp slanted-roof, CdTe, panel, mounted
-    reference product: electricity production, photovoltaic, at 3kWp slanted-roof, CdTe, panel, mounted
+    name: electricity production, photovoltaic, photovoltaic slanted-roof installation, 3 kWp, CdTe, laminated, integrated
+    reference product: electricity, low voltage
     unit: kilowatt hour
     share:
       2020:
@@ -47,7 +67,7 @@ PV:
         maximum: 0.2500
         uncertainty_type: uniform
   CIGS:
-    name: electricity production, photovoltaic, 3kWp slanted-roof installation, CdTe, laminated, integrated
+    name: electricity production, photovoltaic, photovoltaic slanted-roof installation, 3 kWp, CIS, panel, mounted
     reference product: electricity, low voltage
     unit: kilowatt hour
     share:
@@ -69,7 +89,7 @@ PV:
         maximum: 0.0090
         uncertainty_type: uniform
   Perovskite:
-    name: electricity production, photovoltaic, perovskite
+    name: electricity production, photovoltaic, 0.5kWp, perovskite-on-silicon tandem
[…]
