### 2024-07-31 400f09c
Add aggregation of activity categories
