### 2024-07-31 2f78b8a
`run_gsa()` as a standalone function
