+++ dev/timing.py
@@ -1,9 +1,9 @@
-from pathways import Pathways
+from pathways import Pathways, run_gsa
 
 p = Pathways(
     datapackage="remind-SSP2-PkBudg1150-stem-SPS1.zip",
     geography_mapping="geo_mapping_remind.yaml",
-    activities_mapping="act_categories_agg.yaml",
+    #activities_mapping="act_categories_agg.yaml",
 )
 
 vars = [v for v in p.scenarios.coords["variables"].values if v.startswith("FE")]
@@ -17,11 +17,14 @@ p.calculate(
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[2020, 2030, 2040, 2050],
     variables=vars,
-    use_distributions=100,
+    use_distributions=10,
     subshares=True,
-    multiprocessing=True,
 )
 
 p.export_results()
 
-p.run_gsa()
+print(p.lca_results.coords)
+print(p.lca_results.shape)
+print(p.lca_results.sum())
+
+run_gsa()
