+++ pathways/pathways.py
@@ -18,15 +18,11 @@ import xarray as xr
 import yaml
 
 from .data_validation import validate_datapackage
-from .filesystem_constants import DATA_DIR, DIR_CACHED_DB, STATS_DIR, USER_LOGS_DIR
+from .filesystem_constants import DATA_DIR, USER_LOGS_DIR
 from .lca import _calculate_year, get_lca_matrices
 from .lcia import get_lcia_method_names
 from .stats import (
-    create_mapping_sheet,
-    log_results,
-    log_subshares,
-    log_uncertainty_values,
-    run_GSA_delta,
+    log_mc_parameters_to_excel,
 )
 from .subshares import generate_samples
 from .utils import (
@@ -122,96 +118,6 @@ def _fill_in_result_array(
     return np.stack(results, axis=2)
 
 
-def log_mc_parameters_to_excel(
-    model: str,
-    scenario: str,
-    year: int,
-    methods: list,
-    result: dict,
-    uncertainty_parameters: dict,
-    uncertainty_values: dict,
-    tehnosphere_indices: dict,
-    iteration_results: dict,
-    shares: dict = None,
-):
-    export_path = STATS_DIR / f"{model}_{scenario}_{year}.xlsx"
-
-    # create Excel workbook using openpyxl
-    with pd.ExcelWriter(export_path, engine="openpyxl") as writer:
-
-        df_sum_impacts = pd.DataFrame()
-        df_uncertainty_values = pd.DataFrame()
-        df_technology_shares = pd.DataFrame()
-        writer.book.create_sheet("Indices mapping")
-        writer.book.create_sheet("Monte Carlo values")
-        writer.book.create_sheet("Technology shares")
-        writer.book.create_sheet("Total impacts")
-
-        for region, data in result.items():
-
-            total_impacts = np.sum(iteration_results[region], axis=(0, 2, 3))
-
-            df_sum_impacts = pd.concat(
-                [
-                    df_sum_impacts,
-                    log_results(
-                        total_impacts=total_impacts,
-                        methods=methods,
-                        region=region,
-                    ),
-                ]
-            )
-
-            uncertainty_indices = uncertainty_parameters[region]
-            uncertainty_vals = uncertainty_values[region]
-
-            df_uncertainty_values = pd.concat(
-                [
-                    df_uncertainty_values,
-                    log_uncertainty_values(
-                        region=region,
-                        uncertainty_indices=uncertainty_indices,
-                        uncertainty_values=uncertainty_vals,
-                    ),
-                ],
-            )
-
-            if shares:
-                sub_shares = {}
-                for k, v in shares.items():
-                    for x, y in v.items():
-                        if x == year:
-                            for z, w in y.items():
-                                sub_shares[f"{k} - {z}"] = w
-
-                df_technology_shares = pd.concat(
-                    [
-                        df_technology_shares,
-                        log_subshares(
-                            shares=sub_shares,
-                            region=region,
-                        ),
-                    ],
-                )
-
-        indices = tehnosphere_indices[region]
-
-        df_technosphere_indices = create_mapping_sheet(indices=indices)
-
-        df_sum_impacts.to_excel(writer, sheet_name="Total impacts", index=False)
-        df_uncertainty_values.to_excel(
-            writer, sheet_name="Monte Carlo values", index=False
-        )
-        df_technology_shares.to_excel(
-            writer, sheet_name="Technology shares", index=False
-        )
-        df_technosphere_indices.to_excel(
-            writer, sheet_name="Indices mapping", index=False
-        )
-
-        print(f"Monte Carlo parameters added to: {export_path.resolve()}")
-
-
 class Pathways:
[…]
