### 2024-07-31 66e051c
Fix issues with disordered activities categories and locations
