+++ dev/timing.py
@@ -3,7 +3,7 @@ from pathways import Pathways, run_gsa
 p = Pathways(
     datapackage="remind-SSP2-PkBudg1150-stem-SPS1.zip",
     geography_mapping="geo_mapping_remind.yaml",
-    #activities_mapping="act_categories_agg.yaml",
+    activities_mapping="act_categories_agg.yaml",
 )
 
 vars = [v for v in p.scenarios.coords["variables"].values if v.startswith("FE")]
@@ -15,9 +15,14 @@ p.calculate(
     ],
     regions=["CH"],
     scenarios=p.scenarios.pathway.values.tolist(),
-    years=[2020, 2030, 2040, 2050],
+    years=[
+        2020,
+        #2030,
+        #2040,
+        #2050
+    ],
     variables=vars,
-    use_distributions=10,
+    use_distributions=20,
     subshares=True,
 )
 

[…]
