+++ dev/timing.py
@@ -17,9 +17,9 @@ p.calculate(
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[
         2020,
-        #2030,
-        #2040,
-        #2050
+        # 2030,
+        # 2040,
+        # 2050
     ],
     variables=vars,
     use_distributions=20,
