### 2024-08-02 0a32ef6
Fix ref in paper.md
