### 2024-08-05 #6814 @fraukewiese
@ClaraBuettner, @@kcerva, @mkvdhulst
– would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

The submission under review here is [pathways: life cycle assessment of energy transition](https://github.com/polca/pathways)
