### 2024-08-09 6db2a88
Fix subshares iterations
