### 2024-08-09 9a621c9
Add GSA paper material
