+++ dev/GSA Paper/act_categories_agg.yaml
@@ -1,0 +1,158 @@
+Aquaculture: Aquaculture
+Architectural and engineering activities and related technical consultancy: Architectural and Engineering Services
+Building of ships and floating structures: Shipbuilding and Marine Engineering
+Carbon capture: Carbon Management
+Carbon storage: Carbon Management
+Casting of iron and steel: Casting and Metalworking
+Casting of nonferrous metals: Casting and Metalworking
+Collection of nonhazardous waste: Waste Collection and Treatment
+Construction of buildings: Construction and Civil Engineering
+Construction of other civil engineering projects: Construction and Civil Engineering
+Construction of roads and railways: Construction and Civil Engineering
+Construction of utility projects: Construction and Civil Engineering
+Cutting shaping and finishing of stone: Stone and Mineral Processing
+Electric power generation transmission and distribution: Electric Power and Utilities
+Electricity gas steam and air conditioning supply: Electric Power and Utilities
+Extraction of crude petroleum: Petroleum and Natural Gas Extraction
+Extraction of crude petroleum and natural gas: Petroleum and Natural Gas Extraction
+Extraction of natural gas: Petroleum and Natural Gas Extraction
+Extraction of peat: Salt and Other Mineral Extraction
+Extraction of salt: Salt and Other Mineral Extraction
+Finishing of textiles: Textile Manufacturing
+Forging pressing stamping and rollforming of metal powder metallurgy: Metalworking and Forging
+Freight air transport: Freight Transport
+Freight rail transport: Freight Transport
+Freight transport by road: Freight Transport
+Freshwater aquaculture: Aquaculture
+Gathering of nonwood forest products: Agriculture and Crop Production
+Growing of beverage crops: Agriculture and Crop Production
+Growing of cereals except rice leguminous crops and oil seeds: Agriculture and Crop Production
+Growing of citrus fruits: Agriculture and Crop Production
+Growing of fibre crops: Agriculture and Crop Production
+Growing of grapes: Agriculture and Crop Production
+Growing of oleaginous fruits: Agriculture and Crop Production
+Growing of other nonperennial crops: Agriculture and Crop Production
+Growing of other perennial crops: Agriculture and Crop Production
+Growing of other tree and bush fruits and nuts: Agriculture and Crop Production
+Growing of pome fruits and stone fruits: Agriculture and Crop Production
+Growing of rice: Agriculture and Crop Production
+Growing of spices aromatic drug and pharmaceutical crops: Agriculture and Crop Production
+Growing of sugar cane: Agriculture and Crop Production
+Growing of tropical and subtropical fruits: Agriculture and Crop Production
+Growing of vegetables and melons roots and tubers: Agriculture and Crop Production
+Inland freight water transport: Freight Transport
+Logging: Forestry and Logging
+Maintenance and repair of motor vehicles: Transport Equipment Maintenance and Repair
+Manufacture of agricultural and forestry machinery: Automotive and Machinery Manufacturing
+Manufacture of air and spacecraft and related machinery: Automotive and Machinery Manufacturing
+Manufacture of articles of concrete cement and plaster: Other Manufacturing
+Manufacture of basic chemical: Chemicals and Fertilizers
+Manufacture of basic chemicals: Chemicals and Fertilizers
+Manufacture of basic chemicals fertilizers and nitrogen compounds plastics: Chemicals and Fertilizers
+Manufacture of basic iron and steel: Iron and Steel Manufacturing
+Manufacture of basic precious and other nonferrous metals: Iron and Steel Manufacturing
+Manufacture of bicycles and invalid carriages: Automotive and Machinery Manufacturing
+Manufacture of bodies coachwork for motor vehicles manufacture of traile: Automotive and Machinery Manufacturing
[…]
