+++ requirements.txt
@@ -7,8 +7,6 @@ premise
 pyyaml
 bw_processing
 bw2calc >= 2.0.dev18
-cvxopt
-scikit-umfpack
 datapackage
 pyprind
 platformdirs
