+++ .github/workflows/main.yml
@@ -114,7 +114,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest"]
-        python-version: [ "3.10" ]
+        python-version: [ "3.11" ]
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
@@ -165,7 +165,7 @@ jobs:
       fail-fast: false
       matrix:
         os: ["ubuntu-latest"]
-        python-version: [ "3.10" ]
+        python-version: [ "3.11" ]
     steps:
       - uses: actions/checkout@v4
       - uses: conda-incubator/setup-miniconda@v3
@@ -186,7 +186,6 @@ jobs:
           conda update pathways
           conda install pytest
           pytest
-        
 
       - name: Pull distribution 📦 from Anaconda and test it if non-master branch
         if: github.ref != 'refs/heads/master'
