+++ .github/workflows/main.yml
@@ -120,7 +120,7 @@ jobs:
       - uses: conda-incubator/setup-miniconda@v3
         with:
           python-version: ${{ matrix.python-version }}
-          channels: conda-forge,cmutel,konstantinstadler,haasad,pascallesage,romainsacchi
+          channels: conda-forge,cmutel,konstantinstadler,haasad,romainsacchi
           allow-softlinks: true
           auto-update-conda: true
           activate-environment: test
@@ -144,18 +144,14 @@ jobs:
           USER=romainsacchi
           conda install anaconda-client
           conda install conda-build
-          conda install conda-verify
-          conda install pytest
-          conda clean --all
-          conda update conda
           mkdir ~/conda-bld
           conda config --set anaconda_upload no
           export CONDA_BLD_PATH=~/conda-bld
           export VERSION=`date +%Y.%m.%d`
           conda build "conda/" --old-build-string
-          conda-verify $CONDA_BLD_PATH/noarch/$PKG_NAME-$VERSION-py_0.tar.bz2
           ls $CONDA_BLD_PATH/noarch/
           anaconda -t ${{ secrets.ANACONDA_CLOUD }} upload -u $USER $CONDA_BLD_PATH/noarch/$PKG_NAME-$VERSION-py_0.tar.bz2 --force
+
   testconda:
     name: (${{ matrix.python-version }}, ${{ matrix.os }})
     runs-on: ${{ matrix.os }}
