+++ pathways/data/activities_classifications.yaml
@@ -76526,3 +76526,1675 @@
 - carbon dioxide, captured at cement plant
 - kilogram
 : Carbon capture
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, without flue gas condensation, with electricity and heat recovery
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, with flue gas condensation, with electricity and heat recovery
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with wet air pollution control,
+  without flue gas condensation, with electricity and heat recovery
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with wet air pollution control,
+  with flue gas condensation, with electricity and heat recovery
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, with flue gas condensation, with electricity recovery only, with carbon
+  capture and storage
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, with flue gas condensation, with electricity and heat recovery, with
+  carbon capture and storage
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with wet air pollution control,
+  with flue gas condensation, with electricity and heat recovery, with carbon capture
+  and storage
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with wet air pollution control,
+  with flue gas condensation, with electricity and optimized heat recovery, with
+  carbon capture and storage
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, without flue gas condensation, with electricity recovery only, energy
+  allocation
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, without flue gas condensation, with electricity recovery only, energy
+  allocation
+- electricity, medium voltage
+- kilowatt hour
+: Electric power generation transmission and distribution
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, without flue gas condensation, with electricity and heat recovery, energy
+  allocation
+- municipal solid waste
+- kilogram
+: Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, without flue gas condensation, with electricity and heat recovery, energy
+  allocation
+- electricity, medium voltage
+- kilowatt hour
+: Electric power generation transmission and distribution
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
+  control, without flue gas condensation, with electricity and heat recovery, energy
+  allocation
+- heat, district or industrial, other than natural gas
+- megajoule
+: Steam and air conditioning supply
+? !!python/tuple
+- municipal waste incineration, at incineration plant, with semi-dry air pollution
[…]
