+++ dev/timing.py
@@ -16,7 +16,7 @@ p.calculate(
     regions=["CH"],
     scenarios=p.scenarios.pathway.values.tolist(),
     years=[
-        #2020,
+        # 2020,
         # 2030,
         # 2040,
         2050
