### 2024-08-11 db2cff0
Allow running shares without other uncertainties. Also, allow for a MC seed for reproducibility.
