+++ pathways/subshares.py
@@ -268,7 +268,7 @@ def default_dict_factory():
     return defaultdict(dict)
 
 
-def load_and_normalize_shares(ranges: dict, iterations: int) -> dict:
+def load_and_normalize_shares(ranges: dict, iterations: int, seed: int) -> dict:
     """
     Load and normalize shares for parameters to sum to 1 while respecting their specified ranges.
     :param ranges: A dictionary with categories, technologies and market shares data.
@@ -282,7 +282,10 @@ def load_and_normalize_shares(ranges: dict, iterations: int) -> dict:
         for technology, params in technologies.items():
             for y, share in params["share"].items():
                 uncertainty_base = UncertaintyBase.from_dicts(share)
-                random_generator = MCRandomNumberGenerator(uncertainty_base)
+                random_generator = MCRandomNumberGenerator(
+                    params=uncertainty_base,
+                    seed=seed
+                )
                 shares[technology_group][y][technology] = np.squeeze(
                     np.array([random_generator.next() for _ in range(iterations)])
                 )
@@ -375,7 +378,7 @@ def interpolate_for_year(
         shares[technology_group][target_year][technology] = f(target_year)
 
 
-def generate_samples(years: list, iterations: int = 10) -> dict:
+def generate_samples(years: list, iterations: int = 10, seed: int = None) -> dict:
     """
     Generates and adjusts randomly selected shares for parameters to sum to 1
     while respecting their specified ranges, and interpolates missing years.
@@ -385,6 +388,6 @@ def generate_samples(years: list, iterations: int = 10) -> dict:
     :return: A dict with adjusted and interpolated shares for each technology and year.
     """
     ranges = load_subshares()
-    shares = load_and_normalize_shares(ranges, iterations)
+    shares = load_and_normalize_shares(ranges, iterations, seed)
     interpolate_shares(shares, years)
     return shares
