+++ pathways/lca.py
@@ -493,7 +493,7 @@ def _calculate_year(args: tuple):
         shares,
         uncertain_parameters,
         remove_uncertainty,
-        seed
+        seed,
     ) = args
 
     print(f"------ Calculating LCA results for {year}...")
@@ -604,7 +604,7 @@ def _calculate_year(args: tuple):
                 bw_datapackage,
             ],
             use_distributions=True if use_distributions > 0 else False,
-            seed_override=seed
+            seed_override=seed,
         )
 
         with CustomFilter("(almost) singular matrix"):
