### 2024-08-11 4db6f2e
Allow running shares without other uncertainties. Also, allow for a MC seed for reproducibility.
