+++ dev/variables.yaml
@@ -1,0 +1,1 @@
+DIR_CACHED_DB: /data/user/sacchi_r/cache
