### 2024-08-12 b9b33b6
Remove seed from shares
