+++ pathways/subshares.py
@@ -268,7 +268,10 @@ def default_dict_factory():
     return defaultdict(dict)
 
 
-def load_and_normalize_shares(ranges: dict, iterations: int, ) -> dict:
+def load_and_normalize_shares(
+    ranges: dict,
+    iterations: int,
+) -> dict:
     """
     Load and normalize shares for parameters to sum to 1 while respecting their specified ranges.
     :param ranges: A dictionary with categories, technologies and market shares data.
@@ -377,7 +380,10 @@ def interpolate_for_year(
         shares[technology_group][target_year][technology] = f(target_year)
 
 
-def generate_samples(years: list, iterations: int = 10, ) -> dict:
+def generate_samples(
+    years: list,
+    iterations: int = 10,
+) -> dict:
     """
     Generates and adjusts randomly selected shares for parameters to sum to 1
     while respecting their specified ranges, and interpolates missing years.
@@ -387,6 +393,9 @@ def generate_samples(years: list, iterations: int = 10, ) -> dict:
     :return: A dict with adjusted and interpolated shares for each technology and year.
     """
     ranges = load_subshares()
-    shares = load_and_normalize_shares(ranges, iterations,)
+    shares = load_and_normalize_shares(
+        ranges,
+        iterations,
+    )
     interpolate_shares(shares, years)
     return shares
