### 2024-08-24 8bf7a0b
Fix issue about variable-results mapping
