+++ pathways/data/activities_classifications.yaml
@@ -78198,3 +78198,4079 @@
 - municipal solid waste
 - kilogram
 : Treatment and disposal of nonhazardous waste
+? !!python/tuple
+- metal organic vapor phase epitaxy (MOVPE), for GaAs photovoltaic panel
+- gaAs wafer with nanowires
+- square meter
+: Manufacture of batteries and accumulators
+? !!python/tuple
+- perovskite photovoltaic module assembly
+- perovskite photovoltaic module
+- square meter
+: Manufacture of batteries and accumulators
+? !!python/tuple
+- photovoltaic panel installation, 0.5kWp, perovskite
+- photovoltaic panel, perovskite
+- unit
+: Manufacture of batteries and accumulators
+? !!python/tuple
+- hydrogen supply, hydrogen, from petroleum, by truck, as gaseous, over 500 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from petroleum, by truck, as liquid, over 500 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from petroleum, by ship, as liquid, over 2000 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from petroleum, by H2 pipeline, as gaseous, over 500
+  km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, PEM, by truck, as gaseous, over
+  500 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, PEM, by truck, as liquid, over 500
+  km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, PEM, by ship, as liquid, over 2000
+  km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, PEM, by H2 pipeline, as gaseous,
+  over 500 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, AEC, by truck, as gaseous, over
+  500 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, AEC, by truck, as liquid, over 500
+  km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, AEC, by ship, as liquid, over 2000
+  km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, AEC, by H2 pipeline, as gaseous,
+  over 500 km
+- hydrogen, 700 bar
+- kilogram
+: Manufacture of synthetic fuels
+? !!python/tuple
+- hydrogen supply, hydrogen, from electrolysis, SOEC, by truck, as gaseous, over
[…]
