+++ pathways/pathways.py
@@ -389,8 +389,7 @@ class Pathways:
         # remove keys that are not
         # in self.scenarios.variable.values
         self.mapping = {
-            k: self.mapping[k]
-            for k in self.scenarios.coords["variables"].values
+            k: self.mapping[k] for k in self.scenarios.coords["variables"].values
         }
 
         try:
@@ -516,7 +515,9 @@ class Pathways:
             for coords, values in results.items():
                 model, scenario, year = coords
 
-                logging.info(f"Variables in lca_results: {self.lca_results.coords['variable'].values}")
+                logging.info(
+                    f"Variables in lca_results: {self.lca_results.coords['variable'].values}"
+                )
 
                 self.lca_results.loc[
                     dict(
