### 2024-09-10 #6814 @fraukewiese
@rogersamso , @mfastudillo
– would one of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under review here is [pathways: life cycle assessment of energy transition](https://github.com/polca/pathways)
