### 2024-10-01 #6814 @fraukewiese
> @kcerva : Thank you very much! The goal is to finish a review within 6 weeks, including the iteration steps of the reviewers making improvement suggestions and the authors reacting to those and adapting code and paper. This can also already start step-by-step during the process of reviewing. So your review does not have to be finished before already suggestion points for improvement to the authors or questions you might have in the review threat. However, this might take longer that the 6 weeks, so probably 2-3 months is more realistic, also because we still need to find a second reviewer. Thus, I will make you reviewer now but the actual review will start as soon as we have found a second reviewer and I will notify you here in this thread.

@kcerva : I have mixed the the answer from @marc-vdm up with you, that's why I made you a reviewer already. Sorry for that! However, would you still be willing to review this submission taking into consideration my answer about the deadlines? :)
