### 2024-10-02 #7309 @editorialbot
[REVIEW]: pathways: life cycle assessment of energy transition scenarios
**Submitting author:** @romainsacchi (<a href="http://orcid.org/0000-0003-1440-0905">Romain Sacchi</a>)
**Repository:** https://github.com/polca/pathways
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
