### 2024-10-02 #7309 @fraukewiese
@romainsacchi : Could you check the invalid DOI (see above)?
