+++ paper.bib
@@ -165,7 +165,7 @@ year = {2018}
 @article{BORGONOVO2007771,
 abstract = {Uncertainty in parameters is present in many risk assessment problems and leads to uncertainty in model predictions. In this work, we introduce a global sensitivity indicator which looks at the influence of input uncertainty on the entire output distribution without reference to a specific moment of the output (moment independence) and which can be defined also in the presence of correlations among the parameters. We discuss its mathematical properties and highlight the differences between the present indicator, variance-based uncertainty importance measures and a moment independent sensitivity indicator previously introduced in the literature. Numerical results are discussed with application to the probabilistic risk assessment model on which Iman [A matrix-based approach to uncertainty and sensitivity analysis for fault trees. Risk Anal 1987;7(1):22–33] first introduced uncertainty importance measures.},
 author = {Borgonovo, E},
-doi = {https://doi.org/10.1016/j.ress.2006.04.015},
+doi = {10.1016/j.ress.2006.04.015},
 issn = {0951-8320},
 journal = {Reliability Engineering & System Safety},
 keywords = { Global sensitivity analysis, Probabilistic risk assessment, Uncertainty analysis, Uncertainty importance measures,Importance measures},
