+++ .github/workflows/main.yml
@@ -38,6 +38,13 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - name: Install system dependencies for pyenv
+        run: |
+          sudo apt-get update
+          sudo apt-get install build-essential zlib1g-dev libffi-dev libssl-dev \
+                               libbz2-dev libreadline-dev libsqlite3-dev liblzma-dev \
+                               libncurses-dev tk-dev
+
       - name: Install python version
         uses: gabrielfalcao/pyenv-action@v13
         with:
