### 2024-10-15 #7309 @kcerva
@romainsacchi I'm slowly checking all the points. However, perhaps it's useful if I start commenting on some things. This way you can start considering them and address them whenever you're ready.

1. The literature cited is mostly related to LCA applications but it ignores software or tool contributions to the literature that are based on LCA data and use it for further analysis. Please see examples by the Industrial Ecology community and their recent [Industrial Ecology Data Commons initiative](https://www.database.industrialecology.uni-freiburg.de/) and even some of our own work in the [Mat-dp](https://github.com/Mat-dp) model and [how it can be used](https://www.sciencedirect.com/science/article/pii/S0921344924003975) (which I’m mentioning not to try to get it cited but to show examples of what I mean when I say that LCA data is used to underpin other tools or analyses in an automated way and it would be worth reviewing them further).
2. Is there a way to check for double counting in the LCA data and how this affects the environmental impacts that result from the analysis? For example, I’ve seen LCA data in mining processes have issues related to different system boundaries so it has been only by understanding the LCA boundaries in detail that these shortcomings may be addressed. Would this be required in the LCA data included here?
3. I noticed the following comment in this path ([pathways](https://github.com/polca/pathways/tree/main)/[pathways](https://github.com/polca/pathways/tree/main/pathways)/[data](https://github.com/polca/pathways/tree/main/pathways/data)/technologies_shares.yaml): "Check if IAMs make the distinction between onshore and offshore." I can confirm that IAMs do make the distinction between these. In case it's helpful.
