### 2024-10-17 e93c807
Comments from kcerva #1 and #3
