+++ paper.bib
@@ -176,3 +176,37 @@ url = {https://www.sciencedirect.com/science/article/pii/S0951832006000883},
 volume = {92},
 year = {2007}
 }
+
+@article{RECC:2021,
+author = {Pauliuk, S. and Fishman, T. and Heeren, N. and Berrill, P. and Tu, Q. and Wolfram, P. and Hertwich, E.G.},
+doi = {10.1111/jiec.13023},
+journal = {Journal of Industrial Ecology},
+pages = {260--273},
+title = {{Linking service provision to material cycles: A new framework for studying the resource efficiency–climate change (RECC) nexus}},
+url = {https://onlinelibrary.wiley.com/doi/full/10.1111/jiec.13023},
+volume = {25},
+year = {2021}
+}
+
+@article{Mat-dp:2022,
+author = {Barron, K.C. and Cullen, J.M.},
+doi = {10.21105/joss.04460},
+journal = {Journal of Open Source Software},
+pages = {4460},
+title = {{Mat-dp: An open-source Python model for analysing material demand projections and their environmental implications, which result from building low-carbon systems.}},
+url = {https://joss.theoj.org/papers/10.21105/joss.04460},
+volume = {7},
+number = {76},
+year = {2022}
+}
+
+@article{Mat-dp:2024,
+author = {Barron, K.C. and Cullen, J.M.},
+doi = {10.1016/j.resconrec.2024.107803},
+journal = {Resources, Conservation and Recycling},
+pages = {107803},
+title = {{Using open-source tools to project bulk and critical material demand and assess implications for low-carbon energy and transport systems: Introducing Mat-dp model tools.}},
+url = {https://joss.theoj.org/papers/10.21105/joss.04460},
+volume = {209},
+year = {2024}
+}
