+++ paper.bib
@@ -189,7 +189,7 @@ year = {2021}
 }
 
 @article{Mat-dp:2022,
-author = {Barron, K.C. and Cullen, J.M.},
+author = {Cervantes Barron, K. and Cullen, J.M.},
 doi = {10.21105/joss.04460},
 journal = {Journal of Open Source Software},
 pages = {4460},
@@ -201,7 +201,7 @@ year = {2022}
 }
 
 @article{Mat-dp:2024,
-author = {Barron, K.C. and Cullen, J.M.},
+author = {Cervantes Barron, K. and Cullen, J.M.},
 doi = {10.1016/j.resconrec.2024.107803},
 journal = {Resources, Conservation and Recycling},
 pages = {107803},
