### 2024-10-18 1962e78
Extend double-counting paragraph
