### 2024-10-18 #7309 @romainsacchi
Hello @kcerva, regarding your second comment:

The current approach to addressing double counting in pathways relies on the practitioner’s understanding of what the ESM or IAM is modeling to avoid overlaps in the supply chain. For example, when modeling the production of chromite ore concentrate, the inventory in ecoinvent includes inputs such as electricity, heat, diesel burned in machinery, and non-sulfidic tailings treatment.

Since this chromite will also be required upstream in electricity production, if the IAM models the entire electricity generation system for the relevant region, including the electricity demand from this inventory would result in double counting. If the IAM accounts for both heating and electricity, these flows would be doubly counted. Additionally, the electricity required for diesel supply and tailings treatment must be excluded to prevent redundancy.

Therefore, practitioners must manually identify and exclude overlapping activities based on their understanding of the scenario. This approach offers flexibility in canceling out specific groups of activities. For instance, if the scenario also models steel production, the practitioner can adjust the modular approach to exclude the double counting of upstream steel demand.

Quantifying the exact impact of double counting is complex and would require a dedicated case study with specific input data and assumptions, which could form the basis of a separate paper. However, the significance of double counting is clear. For example, in producing 1 kg of chromite concentrate, electricity supply accounts for 68% of the global warming impacts. Double counting this electricity would significantly distort the overall environmental impact assessment. In a quick exercise with the IMAGE model (scenario SSP2-RCP26-SPS1), adjusting for upstream electricity in all sources reduced the overall global warming potential (GWP) by around 25%.

![image001](https://github.com/user-attachments/assets/060201e9-4552-4b16-ac3a-9c0cb7c8a7bd)

In general, quantities provided by the IAM—such as total transport demand or electricity production—should be respected and not adjusted based on the supply chain examination. This ensures that the core scenario assumptions remain intact.

We made an addition to the text (Commit https://github.com/polca/pathways/commit/1962e78316939df2bc8ac4e70b0a320fd3055994) to adress this concern.
