### 2024-10-29 polca/pathways#5 @marc-vdm
Install problems
related to https://github.com/openjournals/joss-reviews/issues/7309

___
When I `conda create -n pw pip` and `pip git+https://github.com/polca/pathways.git` I get:

`ERROR: unknown command "git+https://github.com/polca/pathways.git"`

- [x] Please add proper requirements for pip install

___
When I `conda create -n pw pip` and
```bash
git clone https://github.com/polca/pathways.git
cd pathways
pip install -r requirements.txt
```
- [x] Please note that `cd` is a MacOS and Linux command, I don't expect this to work on Windows (I think it's `dir`?)

I get:

```bash

Cloning into 'pathways'...
remote: Enumerating objects: 2019, done.
remote: Total 2019 (delta 239), reused 243 (delta 224), pack-reused 1743 (from 1)
Resolving deltas: 100% (1366/1366), done.
Collecting numpy==1.24.4 (from -r requirements.txt (line 1))
  Downloading numpy-1.24.4.tar.gz (10.9 MB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [32 lines of output]
      Traceback (most recent call last):
        File "/home/marc/.conda/envs/pw/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
          ~~~~^^
        File "/home/marc/.conda/envs/pw/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
                                   ~~~~^^^^^^^^^^^^^^^^^^^^^^^^
        File "/home/marc/.conda/envs/pw/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 112, in get_requires_for_build_wheel
          backend = _build_backend()
        File "/home/marc/.conda/envs/pw/lib/python3.13/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 77, in _build_backend
          obj = import_module(mod_path)
        File "/home/marc/.conda/envs/pw/lib/python3.13/importlib/__init__.py", line 88, in import_module
                 ~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        File "<frozen importlib._bootstrap>", line 1387, in _gcd_import
        File "<frozen importlib._bootstrap>", line 1360, in _find_and_load
        File "<frozen importlib._bootstrap>", line 1310, in _find_and_load_unlocked
        File "<frozen importlib._bootstrap>", line 488, in _call_with_frames_removed
        File "<frozen importlib._bootstrap>", line 1387, in _gcd_import
        File "<frozen importlib._bootstrap>", line 1360, in _find_and_load
        File "<frozen importlib._bootstrap>", line 1331, in _find_and_load_unlocked
        File "<frozen importlib._bootstrap>", line 935, in _load_unlocked
        File "<frozen importlib._bootstrap_external>", line 1022, in exec_module
        File "<frozen importlib._bootstrap>", line 488, in _call_with_frames_removed
        File "/tmp/pip-build-env-hgtzwtot/overlay/lib/python3.13/site-packages/setuptools/__init__.py", line 16, in <module>
          import setuptools.version
        File "/tmp/pip-build-env-hgtzwtot/overlay/lib/python3.13/site-packages/setuptools/version.py", line 1, in <module>
          import pkg_resources
        File "/tmp/pip-build-env-hgtzwtot/overlay/lib/python3.13/site-packages/pkg_resources/__init__.py", line 2172, in <module>
          register_finder(pkgutil.ImpImporter, find_on_path)
                          ^^^^^^^^^^^^^^^^^^^
      AttributeError: module 'pkgutil' has no attribute 'ImpImporter'. Did you mean: 'zipimporter'?
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```

- [x] Please resolve problems

___
When I `conda create -n pw` and `conda install romainsacchi::pathways` I get the error that `bw2calc` is not installed, after that that `datapackage` is not installed, after which I stopped.

- [ ] Please add proper requirements for conda installation.
