### 2024-10-29 polca/pathways#6 @marc-vdm
paper comments
related to https://github.com/openjournals/joss-reviews/issues/7309

Some comments:

### Statement of need
> Performing scenario-wide LCAs with adjusted life cycle inventories ...

What is meant with 'scenario-wide LCA'? Later 'system-wide'. is used, this makes more sense to me.

> ... methodological complexities, such as defining functional units based on IAM outputs and resolving double-counting issues

Please provide examples or references to discussion on this.

### Description
> Some post-processing is done on the inventory matrices, including dealing with double counting.

Refers back to the double counting from above, what is the problem here, and how is it resolved? The rest of the paragraph attempts to explains this, but I think I don't understand the solution.
