### 2024-10-29 1719fe6
Correct installation instructions
