+++ setup.py
@@ -33,7 +33,7 @@ def package_files(directory):
 setup(
     name="pathways",
     version="0.0.1",
-    python_requires=">=3.9",
+    python_requires=">=3.9,<=3.11",
     packages=packages,
     author="Romain Sacchi <romain.sacchi@psi.ch>",
     license=open("LICENSE").read(),
@@ -44,7 +44,9 @@ setup(
         "numpy==1.24.4",
         "pathlib",
         "pandas",
-        "xarray",
+        "xarray<=2024.2.0",
+        "bw2calc >= 2.0.dev18",
+        "bw2data >= 4.0.dev59",
         "scipy",
         "premise",
         "pyyaml",
