+++ requirements.txt
@@ -7,6 +7,7 @@ premise
 pyyaml
 bw_processing
 bw2calc >= 2.0.dev18
+bw2data >= 4.0.dev59
 datapackage
 pyprind
 platformdirs
