### 2024-10-29 polca/pathways#5 @romainsacchi
1. pip install instructions corrected (`install` argument was missing)
2. `cd` is also a correct command to enter a folder on Windows (maybe you meant `dir`/`ls`?)
3. The issue is caused by the requirement `numpy<=1.24.4` that is incompatible with Python 3.12. Hence, I now restricted the package compatibility to 3.9-3.11. We need `numpy<=1.24.4` because of `bw2io`.
4. I fixed the conda package requirements. It should work fine now.
