### 2024-10-29 e57ecf5
Fix excel logging
