+++ pathways/stats.py
@@ -280,9 +280,16 @@ def log_mc_parameters_to_excel(
         mode = "w"
         if_sheet_exists = None  # No need for if_sheet_exists in write mode
 
-    with pd.ExcelWriter(export_path, engine="openpyxl", mode=mode, if_sheet_exists=if_sheet_exists) as writer:
+    with pd.ExcelWriter(
+        export_path, engine="openpyxl", mode=mode, if_sheet_exists=if_sheet_exists
+    ) as writer:
         empty_df = pd.DataFrame()  # Empty DataFrame for clearing purposes
-        for sheet in ["Indices mapping", "Monte Carlo values", "Technology shares", "Total impacts"]:
+        for sheet in [
+            "Indices mapping",
+            "Monte Carlo values",
+            "Technology shares",
+            "Total impacts",
+        ]:
             empty_df.to_excel(writer, sheet_name=sheet, index=False)
 
         df_sum_impacts = pd.DataFrame()
