### 2024-10-29 polca/pathways#5 @marc-vdm
1. I now get the same error as with the `git clone` option
2. yes, my mistake
3. see 1
4. still the same problem as before, but your channel is also not updating pathways: https://anaconda.org/romainsacchi/pathways/files
