### 2024-10-29 polca/pathways#5 @romainsacchi
1 & 3: could it be that the environment uses Python 3.12?
