### 2024-10-29 bed3211
Fix setup.py
