+++ setup.py
@@ -33,7 +33,7 @@ def package_files(directory):
 setup(
     name="pathways",
     version="0.0.1",
-    python_requires=">=3.9,<=3.11",
+    python_requires=">=3.9,<3.12",
     packages=packages,
     author="Romain Sacchi <romain.sacchi@psi.ch>",
     license=open("LICENSE").read(),
