+++ .github/workflows/main.yml
@@ -73,7 +73,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ "ubuntu-latest" ]
-        python-version: [ "3.10" ]
+        python-version: [ "3.11" ]
 
     steps:
       - uses: actions/checkout@v4
