### 2024-10-29 eb6e851
Fix setup.py
