+++ .github/workflows/main.yml
@@ -83,8 +83,6 @@ jobs:
           python -m
           pip install
           build
-          --user
-          --break-system-packages
 
       - name: Build a binary wheel and a source tarball
         run: >-
