### 2024-10-29 polca/pathways#5 @marc-vdm
Yes that resolves 1&3, should have looked better.

For 4: my conda sees the version of today with `conda search -c romainsacchi -f pathways` but still has the original error, also with `python=3.11`
