### 2024-10-29 polca/pathways#5 @romainsacchi
4. Indeed, I had an issue with the pipeline. The anaconda channel is now updated.
