### 2024-10-29 polca/pathways#5 @romainsacchi
4. I just tried `conda install romainsacchi::pathways` which fetches today's release (from 20 minutes ago), and it seems to work fine now (and `bw2calc` and `datapackage` are now part of the environment). Can you confirm?
