### 2024-10-29 polca/pathways#5 @marc-vdm
For me it keeps failing:

```bash
(pw) marc@framework7640u:~$ conda install romainsacchi::pathways=2024.10.29
Channels:
 - defaults
 - romainsacchi
Platform: linux-64
Solving environment: failed

LibMambaUnsatisfiableError: Encountered problems while solving:
  - nothing provides bw2calc needed by pathways-2024.10.29-py_0

Could not solve for environment specs
The following package could not be installed
└─ pathways 2024.10.29*  is not installable because it requires
   └─ bw2calc, which does not exist (perhaps a missing channel).

```
