### 2024-10-29 c311394
Fix conda installation instruction
