### 2024-10-29 polca/pathways#5 @romainsacchi
I see. I think that is because I forgot to specify the `conda-forge` channel. I changed the instruction to:
`conda install -c conda-forge -c romainsacchi pathways`. Now it should find the missing packages.
