### 2024-10-30 02a702f
Add version requirements for bw2calc and bw2data
