+++ conda/meta.yaml
@@ -26,8 +26,8 @@ requirements:
     - premise
     - pyyaml
     - bw_processing
-    - bw2calc
-    - bw2data
+    - bw2calc >= 2.0.dev23
+    - bw2data >= 4.0.dev59
     - datapackage
     - pyprind
     - platformdirs
