### 2024-10-30 dbc3ece
Silence datapackage validation warnings that are due to the presence of the YAML file.
