### 2024-10-30 7bf0c7f
Update conda recipe file
