### 2024-11-01 polca/pathways#6 @marc-vdm
@alvarojhahn There's a space missing at the end of the sentence on line 150.

Besides that LGTM
