### 2024-11-13 #7309 @fraukewiese
@romainsacchi : Congratulations to the positive reviews :)
Two more minor points to improve:
- Could you please check "doi.org/10.1021/acs.est.9b06484 is INVALID because of 'doi.org/' prefix"
- line 58: Could you check the usage of the word "said". Should it be "...characterization of the scenario." - I am not fully sure, it might also be due to my misunderstanding since English is not my native language.
