+++ paper.bib
@@ -139,7 +139,7 @@ year = {2018}
 
 @article{Vandepaer:2020,
   author = {Vandepaer, Laurent and Panos, Evangelos and Bauer, Christian and Amor, Ben},
-  doi = {doi.org/10.1021/acs.est.9b06484},
+  doi = {10.1021/acs.est.9b06484},
   journal = {Environmental Science & Technology},
   pages = {5081-5092},
   title = {{Energy system pathways with low environmental impacts and limited costs: Minimizing climate change impacts produces environmental cobenefits and challenges in toxicity and metal depletion categories}},
