### 2024-11-13 f0e581e
add double counting
