### 2024-11-13 3e20c2d
add smart categories
