+++ pathways/data/smart_categories.yaml
@@ -1,0 +1,317 @@
+Energy:
+  Electricity:
+    Industrial:
+      PV:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, photovoltaic
+          name_mask:
+            - facade installation
+            - flat-roof installation
+            - slanted-roof installation
+            - electricity production, photovoltaic, at 3kWp slanted-roof, CdTe, panel, mounted
+            - electricity production, photovoltaic, at 3kWp slanted-roof, CIS, laminated, integrated
+#            - commercial
+#            - residential
+          product_fltr:
+            - electricity
+      Hydro:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, hydro
+          product_fltr:
+            - electricity, high voltage
+      Oil:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, at co-generation oil
+            - electricity production, at oil-fired
+            - electricity production, oil
+            - heat and power co-generation, oil
+          product_fltr:
+            - electricity, high voltage
+      Gas:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, at co-generation natural gas-fired
+            - electricity production, at natural gas-fired
+            - electricity production, natural gas
+            - natural gas, burned in
+            - electricity production, from combined cycle plant, synthetic natural gas
+            - heat and power co-generation, natural gas
+          product_fltr:
+            - electricity, high voltage
+            - electricity, low voltage
+      Biogas:
+        ecoinvent_aliases:
+          name_fltr:
+            - biomethane, low pressure burned
+            - heat and power co-generation, biogas
+          product_fltr:
+            - electricity
+      Nuclear:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, nuclear
+          product_fltr:
+            - electricity
+      Waste2Grid:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity, from municipal waste
+            - treatment of
+          product_fltr:
+            - electricity,
+          product_mask:
+            - electricity, for reuse in municipal waste incineration only
+      Waste2Treatment:
+        ecoinvent_aliases:
+          name_fltr:
+            - treatment of
+          product_fltr:
+            - electricity, for reuse in municipal waste incineration only
+      Wind:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, wind
+          product_fltr:
+            - electricity
+      Geothermal:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, deep geothermal
+          product_fltr:
+            - electricity
+      Hydrogen:
+        ecoinvent_aliases:
+          name_fltr:
+            - electricity production, at hydrogen-fired
+            - electricity production, from hydrogen-fired
+          product_fltr:
+            - electricity
+      Biomass:
+        ecoinvent_aliases:
+          name_fltr:
+            - biomass-fired
+            - co-firing wood
+            - wood-fired
+            - wood
+            - paraná pine
+            - hardwood
+          name_mask:
+            - treatment
+          product_fltr:
+            - electricity
+      Battery:
+        ecoinvent_aliases:
+          name_fltr:
+            - battery
+          product_fltr:
+            - electricity
+      CSP:
+        ecoinvent_aliases:
+          name_fltr:
+            - solar
+            - parabolic trough
+            - 50 MW
+          product_fltr:
+            - electricity
+      CAES:
+        ecoinvent_aliases:
+          name_fltr:
+            - compressed air energy storage
+          product_fltr:
+            - electricity
+      Import:
+        ecoinvent_aliases:
+          name_fltr:
+            - import
+          product_fltr:
[…]
