+++ pathways/__init__.py
@@ -1,4 +1,4 @@
-__version__ = (0, 0, 1)
+__version__ = (1, 0, 0)
 __all__ = ("__version__", "Pathways", "run_gsa")
 
 
