### 2024-11-13 #7309 @romainsacchi
Hello @fraukewiese, thank you very much!

I released the software under the tag `1.0.0`.

The software is now archived under that version on Zenodo, with the doi: [10.5281/zenodo.14155224](https://doi.org/10.5281/zenodo.14155224)
