### 2024-11-13 #7309 @romainsacchi
@fraukewiese: the minor points have been addressed in the manuscript as well.
