### 2024-11-18 #7309 @romainsacchi
Thank you very much @fraukewiese. And also thank you to the reviewers' team @kcerva @marc-vdm .
I only changed the wording in the Acknowledgment section.
Aside of this, the manuscript looks ready to me.
