### 2024-11-18 #7309 @kyleniemeyer
Hi @romainsacchi, in your paper, you have a few places with multiple citations that need to be combined into one; please use the `[@author1:2001; @author2:2001]` citation format in those cases.
