### 2024-11-18 #7309 @romainsacchi
Hi @kyleniemeyer, thanks! I think these issues are now fixed.
