### 2024-11-18 #7309 @kyleniemeyer
@romainsacchi can you remove the code field around the software name in the paper? That is resulting in `<code>` in the paper's webpage.
